Converts a R object to another

convert_to converts a R object to another object of a predefined class. It have the mission to facilitate conversions between any kind of R object in a simple and fast way.

This function also supports date/time parsing and value transformations. Check the details section to learn more.

convert_to(x, class, ..., quiet = FALSE)

# S3 method for character
convert_to(
  x,
  class,
  orders = c("HMS", "HM", "H"),
  ...,
  tz = "UTC",
  input_unit = NULL,
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  ignore_date = TRUE,
  quiet = FALSE
)

# S3 method for numeric
convert_to(
  x,
  class,
  orders = c("HMS", "HM", "H"),
  ...,
  tz = "UTC",
  input_unit = NULL,
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  ignore_date = TRUE,
  quiet = FALSE
)

# S3 method for Duration
convert_to(
  x,
  class,
  ...,
  tz = "UTC",
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  quiet = FALSE
)

# S3 method for Period
convert_to(
  x,
  class,
  ...,
  tz = "UTC",
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  quiet = FALSE
)

# S3 method for difftime
convert_to(
  x,
  class,
  ...,
  tz = "UTC",
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  quiet = FALSE
)

# S3 method for hms
convert_to(
  x,
  class,
  ...,
  tz = "UTC",
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  quiet = FALSE
)

# S3 method for Date
convert_to(
  x,
  class,
  ...,
  tz = "UTC",
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  ignore_date = TRUE,
  quiet = FALSE
)

# S3 method for POSIXt
convert_to(
  x,
  class,
  ...,
  tz = "UTC",
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  ignore_date = TRUE,
  quiet = FALSE
)

# S3 method for Interval
convert_to(
  x,
  class,
  ...,
  tz = "UTC",
  output_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears(),
  quiet = FALSE
)

convert_to_tu(x, output_unit, ...)

convert_to_uu(x, input_unit, output_unit, ...)

Arguments

x
Any R object, provided that it has an assigned method.

class
A string indicating the class of the output.

quiet
(optional) A logical value indicating if warnings or messages are allowed with the output (default: FALSE).

orders
(optional) A character vector of date/time formats to parse a x value of class character or numeric (default: c("HMS", "HM", "H")).

tz
(optional) A string indicating the time zone with which to convert/parse x.

input_unit
(optional) A string indicating the unit of x.

output_unit
(optional) A string indicating the desire output unit.

month_length
(optional) A duration object or a non negative numeric value with the number of seconds equivalent to the month length. See topic "Year and month lengths" in the details section to learn more (default: lubridate::dmonths(), which is equivalent to 30.4375 days).

year_length
(optional) A duration object or a non negative numeric value with the number of seconds equivalent to the year length. Ssee topic "Year and month lengths" in the details section to know more (default: lubridate::dyears(), which is equivalent to 365.25 days).

ignore_date
(optional) A logical value indicating if dates must be ignored from Date or POSIXt objects, when converting they to numeric.

Value

A R object of the indicated class.

Details

convert_to was designed to be a simple to use function with powerful applications. However, please note that it may not work for all cases. Test it first before applying it to large datasets.

To learn more about how to handle date and time objects, check chapter "Dates and Times" from Wickham and Grolemund (n.d.) and chapter "Technical Representation of Data" from Loo and Jonge (2018).

Wrappers

convert_to has two wrappers functions for convenience.

convert_to_tu was made to help make transformations from date/time objects to units more easily.

convert_to_tu <- function(x, output_unit, ...)

convert_to_uu was made to help make transformations from unit to unit more easily.

convert_to_uu <- function(x, input_unit, output_unit, ...)

class argument

The class argument indicates the desired class of the output object . At the moment, it accepts the following values: "character", "integer", "double", "numeric", "Duration", "Period", "difftime", "hms", "Date", "POSIXct", and "POSIXlt" (case insensitive).

orders argument

orders is an optional argument to indicate date/time formats for parsing character and numeric objects. This kind of parsing is mainly based on lubridate::parse_date_time(). Please check it documentation to set the right formats for your input (default: c("HMS", "HM", "H")).

input_unit and output_unit arguments

To be able to transform date/time objects and numeric values to decimal time, radians, and degrees, convert_to need to know the input_unit (default: NULL) of x and the output_unit (default: NULL). Transformations of Date/time objects only need the latter.

At the moment, valid values for this two variables are:

"S": for decimal seconds.

"M": for decimal minutes.

"H": for decimal hours.

"d": for decimal days.

"W": for decimal weeks.

"m": for decimal months.

"y": for decimal years.

"date_decimal": for decimal dates

"rad": for radians.

"deg": for degrees.

tz argument

Time zones (tz) are used only if class is set to "Date", "POSIXct", or "POSIXlt". See base::timezones to know more about time zone values (default: "UTC").

Parsing limitations

convert_to uses lubridate::parse_date_time() to convert character and numeric objects to date/time. Since parse_date_time() outputs a POSIXt object, character and numeric inputs cannot have time values equal or greater than 24 hours.

That limits the set of convert_to applications (e.g when you want to parse a character to a duration object of 35 minutes and 30 seconds). To get around this, some exceptions were made to orders equal to "H", "M", "S", HM, or HMS (e.g convert_to(c(10, 45, 100), "duration", "M")). For HM and HMS exceptions, minutes and seconds are limited to [0-59], and, when hours exceeds 2 digits, a : must be allocated between hours and minutes.

Different outputs

Your output can change according to your settings. Here are some examples.

When class = "character"

convert_to will return a base::as.character() output if class is set to "character". When x is parsed for date/time, and there's no indication of year, convert_to will return only a character string with a hms time.

You can also parse a character value and transform it direct to an unit. See examples section to learn more.

When class = "numeric"

convert_to will return a base::as.numeric() output if class is set to "numeric". For dates or date-time objects, the output will be the total of seconds from the UNIX origin date (1970-01-01 00:00:00 UTC) (See Unix time to know more). For time values, the output will be the total of seconds.

The output for class = "numeric" can also be different if input_unit and output_unit are assigned.

Interval objects

Interval objects are treated like difftime objects. That is, Interval objects will be converted to the time difference of the interval.

Round-off errors

This function is not optimized to deal with round-off errors.

Year and month lengths

The length of months and years can vary. For example, March have 31 days, while April have have 30. Due to leap years, the same can be said to year lengths.

To address this problem, convert_to use as default the mean of possible values for months and years, used to calculate month and year durations in the lubridate package (see: lubridate::dmonths() and lubridate::dyears()). If you like, you can reset this assigning other values to the arguments month_length and year_length.

month_length and year_length values must be assigned with the number of seconds equivalent to unit duration. You can also assign a lubridate duration object (see lubridate::duration()) if you like.

Decimal dates

convert_to can be used as a wrapper to lubridate::decimal_date() by creating decimal dates that can be used on lubridate::date_decimal().

Decimal dates are a kind of decimal time where the year corresponds to the integer part of the value. The months, days, hours, minutes, and seconds elements are picked so the date-time will represent the fraction of the year expressed by decimal.

Beware that decimal dates are not very precise! It cannot be used as a way to store time data.

References

Van der Loo, M., & De Jonge, E. (2018). Statistical data cleaning with applications in R. Hooboken, NJ: John Wiley & Sons. doi: 10.1002/9781118897126.

Wickham, H, & Grolemund. (n.d.). R for data science. Sebastopol, CA: O'Reilly Media. Retrieved from https://r4ds.had.co.nz.

Examples

## ** conversion of character or numeric values to date/time objects **
convert_to("10:00 PM", "hms", "IMp")

#> 22:00:00

#> [1] 22:00:00 # Expected
convert_to("21:00 AM", "Period")

#> [1] "21H 0M 0S"

#> [1] "21H 0M 0S" # Expected
convert_to("2020-01-01 10:00:00", "Date", "ymd HMS")

#> [1] "2020-01-01"

#> [1] "2020-01-01" # Expected
convert_to(13, "POSIXct", "H")

#> NULL

#> [1] "0000-01-01 13:00:00 UTC" # Expected
convert_to("2020-01-01 12:31:05", "POSIXct", "ymd HMS", tz = "EST")

#> [1] "2020-01-01 12:31:05 EST"

#> [1] "2020-01-01 12:31:05 EST" # Expected
convert_to("03/07/1982 13:00", "POSIXlt", "dmy HM")

#> [1] "1982-07-03 13:00:00 UTC"

#> [1] "1982-07-03 13:00:00 UTC" # Expected

## ** conversion of date/time objects to decimal time, radians or degrees **
convert_to("1015", "numeric", output_unit = "rad")

#> [1] 2.683444

#> [1] 2.683444
convert_to(1.308997, "duration", input_unit = "rad")

#> NULL

#> [1] "18000.0432s (~5 hours)"

## ** conversion between date/time objects **
convert_to(lubridate::duration(120), "hms")

#> 00:02:00

#> 00:02:00 # Expected
convert_to(hms::as_hms("13:45:05"), "POSIXct")

#> [1] "0000-01-01 13:45:05 UTC"

#> [1] "0000-01-01 13:45:05 UTC" # Expected
convert_to(lubridate::period(60), "POSIXct")

#> [1] "0000-01-01 00:01:00 UTC"

#> [1] "0000-01-01 00:01:00 UTC" # Expected
convert_to(as.POSIXct("2020-01-01 12:31:05", tz = "EST"),
                     "POSIXct")

#> [1] "2020-01-01 12:31:05 UTC"

#> [1] "2020-01-01 12:31:05 UTC" # Expected

## ** transformation between numeric units **
convert_to(1.308997, "numeric", input_unit = "rad", output_unit = "H")

#> NULL

#> [1] 5 # Expected