Convert values to date/time

convert_to_date_time converts a date/time value to a predefined object class in a fast and easy way. It can be use to convert/parse objects of class character, numeric, difftime, Duration, hms, Period, Date, POSIXct, POSIXlt and Interval.

Conversions to decimal time and radians are also possible by using other functions of the convert_to family. Check the details section to learn how.

convert_to_date_time(x, class, orders = c("HMS", "HM", "H"), tz = "UTC")

# S3 method for character
convert_to_date_time(x, class, orders = c("HMS", "HM", "H"), tz = "UTC")

# S3 method for numeric
convert_to_date_time(x, class, orders = c("HMS", "HM", "H"), tz = "UTC")

# S3 method for difftime
convert_to_date_time(x, class, orders = c("HMS", "HM", "H"), tz = "UTC")

# S3 method for Duration
convert_to_date_time(x, class, orders = c("HMS", "HM", "H"), tz = "UTC")

# S3 method for hms
convert_to_date_time(x, class, orders = c("HMS", "HM", "H"), tz = "UTC")

# S3 method for Period
convert_to_date_time(x, class, orders = c("HMS", "HM", "H"), tz = "UTC")

# S3 method for Date
convert_to_date_time(x, class, orders = c("HMS", "HM", "H"), tz = "UTC")

# S3 method for POSIXt
convert_to_date_time(x, class, orders = c("HMS", "HM", "H"), tz = "UTC")

# S3 method for Interval
convert_to_date_time(x, class, orders = c("HMS", "HM", "H"), tz = "UTC")

Arguments

x
A character or numeric vector to parse to a date/time object or a date/time object of one of the following classes: difftime, Duration, hms, Period, Date, POSIXct, POSIXlt, and Interval.

class
A character vector indicating the class of the output. Valid values are: "character", "numeric", "difftime", "Duration", "hms", "Period", "Date", "POSIXct" and "POSIXlt" (case insensitive).

orders
A character vector of date/time formats that correspond to x (only needed If x is character or numeric). You can try to parse x trying one or more formats, in order of importance. Check the lubridate lubridate::parse_date_time() function to see the right format values for your input (default: c("HMS", "HM", "H")).

tz
A character string that specifies the time zone with which to convert/parse x. Time zones are used only if class is set to "Date", "POSIXct", or "POSIXlt" (see base::timezones to know more) (default: "UTC").

Value

A date/time object of the indicated class.

Details

convert_to_date_time is heavily based on tidyverse's lubridate::lubridate-package package. Its main objective is to group together several parse methods for the main date/time classes.

This function was made to help users and to assist other functions from the MCTQ package. It may not work for all cases. For better results, make sure that your values fit the argument parameters on the function of your choice.

Orders

This function is mainly based on lubridate's lubridate::parse_date_time() function. If needed, please check it documentation to set the right orders for your input, .

convert_to_date_time is set by default to guess the most common used patterns of data collection for the Munich Chronotype Questionnaire (MCTQ), i.e "HMS", "HM", and "H".

Limitations

convert_to_date_time uses lubridate::parse_date_time() to convert character and numeric objects. Since parse_date_time() output a POSIXt object, character and numeric inputs cannot have time values equal or greater than 24 hours.

That limits the set of convert_to_date_time applications (e.g when you want to parse a character to a duration object of 35 hours and 30 minutes). A exception was made to character and numeric integerish objects with order = "H", order = "M", and order = "S" (e.g convert_to_date_time(c(10, 45, 100), "duration", "M").

To go around this limitation, learn about how to use the lubridate::lubridate-package package.

AM/PM signaling

If your value have an AM/PM signal in your value, convert_to_date_time will try to fix it for you. You can also use the "Op" format on the orders variable to match this indicators (check lubridate::parse_date_time() to know more about it).

Converting date/time objects to decimal time or radians (or the reverse)

convert_to_date_time don't perform conversions for decimal time or radians, but there are easy ways to do it with the convert_to family.

Converting a date/time object to decimal time or radian

Use convert_to_decimal to convert date/time objects to decimal time, or use convert_to_rad to convert date/time objects to radians.

Converting a decimal time to a date/time object

Convert the decimal time to a date/time object using the duration family of functions from the lubridate package. Then, use convert_to_date_time to convert the value to what you which.

EXAMPLE: To convert a decimal time of 23.5 hours to an object of class hms, do:

convert_to_date_time(lubridate::dhours(23.5), "hms")
#> 23:30:00

Converting radians to a date/time object

Convert the radian value to a decimal time using convert_to_decimal. Them, follow the same procedure showed above.

EXAMPLE: To convert 1.309 radian to an object of class duration, do:

convert_to_date_time(lubridate::dhours(convert_to_decimal(1.309)),
"duration")
#> [1] "18000s (~5 hours)"

NAs as output

convert_to_date_time will return a NA value when were not possible to convert the input to the class assigned (sorry!).

Outputs

Your output can change according to the class value indicated on class. Here are some examples.

When class = "character"

convert_to_date_time will return a base::as.character() output if class is set to "character". If there's no indication of year in x, convert_to_date_time will return only a character string with a hms time.

When class = "numeric"

convert_to_date_time will return a base::as.numeric() output if class is set to "numeric". For dates or date-time objects, the output is the total of seconds from the UNIX origin date (1970-01-01 00:00:00 UTC) (See Unix time to know more). For time values, the output is the total of seconds.

Interval objects

Class interval objects are treated like difftime objects. That is, interval objects will be converted to the time difference of the interval.

Note

To do:

Fix conversion for hours equal or greater to 24 for character and numeric variables.

Fix time attribution to Periods (remove functionality?).

Time zone warning

You must be careful when working with different time zones (tz), because some of them can be affected with Daylight Saving Time (DST).

It's unlike that you going to have to worry about time zone while working with the MCTQ. Because of that, convert_to_date_time is set to "UTC" by default for Date, POSIXct, and POSIXlt outputs.

To know more about how to handle date and time objects, check the chapter "Dates and Times" from the book R for Data Science of Hadley Wickham & Garrett Grolemund.

See also

Examples

if (FALSE) {
## ** conversion of character or numeric values to date/time objects **
convert_to_date_time("10:00 PM", "hms")
#> [1] 22:00:00
convert_to_date_time("21:00 AM", "Period")
#> [1] "21H 0M 0S"
convert_to_date_time("2020-01-01 10:00:00", "Date", "ymd HMS")
#> [1] "2020-01-01"
convert_to_date_time(13, "POSIXct", "H")
#> [1] "0000-01-01 13:00:00 UTC"
convert_to_date_time("2020-01-01 12:31:05", "POSIXct", "ymd HMS", tz = "EST")
#> [1] "2020-01-01 12:31:05 EST"
convert_to_date_time("03/07/1982 13:00", "POSIXlt", "dmy HM")
#> [1] "1982-07-03 13:00:00 UTC"

## ** conversion between date/time objects **
convert_to_date_time(lubridate::duration(120), "hms")
#> 00:02:00
convert_to_date_time(hms::as_hms("13:45:05"), "POSIXct")
#> [1] "0000-01-01 13:45:05 UTC"
convert_to_date_time(lubridate::period(60), "POSIXct")
#> [1] "0000-01-01 00:01:00 UTC"
convert_to_date_time(as.POSIXct("2020-01-01 12:31:05", tz = "EST"),
                     "POSIXct")
#> [1] "2020-01-01 12:31:05 UTC"
}