Convert date/time or radians to decimal time

convert_to_decimal converts date/time objects or numeric objects with radian values to decimal time.

convert_to_decimal(
  x,
  unit = "H",
  round = FALSE,
  digits = 3,
  custom_unit = NULL,
  month_length = lubridate::dmonths(),
  year_length = lubridate::dyears()
)

Arguments

x
A date/time object or a numeric vector with radian values. Only the following classes are allowed for date/time objects: difftime, Duration, hms, Period, Date, POSIXct, POSIXlt, and Interval.

unit
A character string indicating the unit of transformation. Go to details section to see the valid values (default: "H").

round
A logical value indicating if the output must be rounded (default: FALSE).

digits
An integer number indicating the number of decimal places or significant digits to be used when round = TRUE (see base::round to know more) (only required when round = TRUE) (default: 3).

custom_unit
A duration object or a non negative numeric value with the number of seconds equivalent to the custom unit (optional) (see topic "Custom unit" in the details section to know more). .

month_length
A duration object or a non negative numeric value with the number of seconds equivalent to the month length (optional) (see topic "Year and month lengths" in the details section to know more) (default: lubridate::dmonths(), which is equivalent to 30.4375 days).

year_length
A duration object or a non negative numeric value with the number of seconds equivalent to the year length (optional) (see topic "Year and month lengths" in the details section to know more)(default: lubridate::dyears(), which is equivalent to 365.25 days).

Value

A numeric vector with decimal times.

Details

unit valid values

convert_to_decimal can do the following transformations. Use the value of your choose in the unit argument.

"H": for decimal hours.

"M": for decimal minutes.

"S": for decimal seconds.

"d": for decimal days.

"W": for decimal weeks.

"m": for decimal months.

"y": for decimal years.

"date_decimal": for decimal dates. See topic "Decimal dates" to know more.

"custom_unit": for custom units. See "Custom unit" topic to know more.

This list of formats correspond to the list used on lubridate::parse_date_time().

Year and month lengths

The length of months and years can vary. For example, March have 31 days, while April have have 30. Due to leap years, the same can be said to year lengths.

To address this problem, convert_to_decimal use as default the mean of possible values for months and years, used to calculate month and year durations on the lubridate package (see: lubridate::dmonths() and lubridate::dyears()). However, if you like, you can reset this values assigning other values to the arguments month_length and year_length.

month_length and year_length values must be assigned with the number of seconds equivalent to the unit duration.

Tip: Use lubridate::duration() to help you assign values (e.g. lubridate::ddays(30)).

Example: If you want to use a month with 30 days, set month_length to lubridate::ddays(30).

Decimal dates

convert_to_decimal can be used as a wrapper to lubridate::decimal_date() by creating decimal dates that can be used on lubridate::date_decimal().

Decimal dates are a kind of decimal time where the year corresponds to the integer part of the value. The months, days, hours, minutes, and seconds elements are picked so the date-time will accurately represent the fraction of the year expressed by decimal.

Beware that decimal dates are not very precise! It cannot be used as a way to store time data.

Example:

x <- lubridate::ymd("2009-02-10")
decimal <- lubridate::decimal_date(x) # 2009.11
lubridate::date_decimal(decimal)
#> [1] "2009-02-10 UTC"

Custom unit

convert_to_decimal can convert time to decimal using any numeric custom unit.

When using a custom unit, the argument custom_unit must be assigned with the number of seconds equivalent to the unit duration. You can also assign a lubridate duration object to custom_unit if you like.

Tip: Use lubridate::duration() to help you assign values (e.g. lubridate::dhours(1)).

Example: if you want to convert a duration of 3 hours to decimal of 1.5 hours, do:

x <- lubridate::dhours(3)
convert_to_decimal(x, unit = "custom_unit",
                   custom_unit = lubridate::dhours(1.5))
#> [1] 2

Radian transformation

convert_to_decimal assumes that any numeric value assigned on x represents a radian value. Just assign it on x and set the unit of your choose.

Example: if you want to convert 1 radian to decimal hours, do:

x <- 1
convert_to_decimal(1)
#> [1] 3.819719

Round-off errors

This function is not optimized to deal with round-off errors. If you need a very precise transformation, convert_to_decimal may not attend you.

Interval objects

Class interval objects used on x are treated like difftime objects, i.e convert_to_decimal will convert the time difference of the interval to a decimal value.

See also

Examples

## ** conversion of date/time objects to decimal time **
x <- hms::as_hms("01:00:00")
convert_to_decimal(x, "S")

#> [1] 3600

#> [1] 3600
convert_to_decimal(x, "M")

#> [1] 60

#> [1] 60
convert_to_decimal(x)

#> [1] 1

#> [1] 1
convert_to_decimal(x, "d")

#> [1] 0.04166667

#> [1] 0.04166667
convert_to_decimal(x, "d", round = TRUE, digits = 3)

#> [1] 0.042

#> [1] 0.042
convert_to_decimal(x, "W", round = TRUE, digits = 3)

#> [1] 0.006

#> [1] 0.006
convert_to_decimal(x, "W")

#> [1] 0.005952381

#> [1] 0.005952381
convert_to_decimal(x, "m")

#> [1] 0.001368925

#> [1] 0.001368925
convert_to_decimal(x, "m", month_length = lubridate::ddays(30.4375))

#> [1] 0.001368925

#> [1] 0.001368925
convert_to_decimal(x, "m", month_length = lubridate::ddays(30))

#> [1] 0.001388889

#> [1] 0.001388889
convert_to_decimal(x, "y")

#> [1] 0.0001140771

#> [1] 0.0001140771
convert_to_decimal(x, "y", year_length = lubridate::ddays(365.25))

#> [1] 0.0001140771

#> [1] 0.0001140771
convert_to_decimal(x, "y", year_length = lubridate::ddays(365))

#> [1] 0.0001141553

#> [1] 0.0001141553
convert_to_decimal(x, "custom_unit", custom_unit = lubridate::dhours())

#> [1] 1

#> [1] 1
convert_to_decimal(x, "custom_unit", custom_unit = 3600)

#> [1] 1

#> [1] 1
convert_to_decimal(x, "custom_unit", custom_unit = 3600 * 2)

#> [1] 0.5

#> [1] 0.5

## ** conversion of date/time objects to decimal dates **
x <- lubridate::ymd("2020-02-01")
convert_to_decimal(x, "date_decimal")

#> [1] 2020.085

#> [1] 2020.085
lubridate::date_decimal(convert_to_decimal(x, "date_decimal"))

#> [1] "2020-01-31 23:59:59 UTC"

#> [1] "2020-02-01 00:00:17 UTC" # Note the imprecision

## ** conversion of radian values to decimal time **
x <- convert_to_rad(hms::as_hms("23:30:30"))
convert_to_rad(hms::as_hms("23:30:30"))

#> [1] 6.154467

#> [1] 6.154467
convert_to_decimal(x, "H")

#> [1] 23.50833

#> [1] 23.5083
convert_to_date_time(lubridate::dhours(convert_to_decimal(x, "H")), "hms")

#> 23:30:30

#> 23:30:30