Convert date/time or decimal hours to radians

convert_to_rad converts date/time objects or numeric objects with decimal hours to radians.

convert_to_rad(x, round = FALSE, digits = 3)

Arguments

x
A date/time object or a numeric vector with decimal hours. Only the following classes are allowed for date/time objects: difftime, Duration, hms, Period, POSIXct, POSIXlt, and Interval.

round
A logical value indicating if the output must be rounded (default: FALSE).

digits
An integer number indicating the number of decimal places or significant digits to be used when round = TRUE (see base::round to know more) (only required when round = TRUE) (default: 3).

Value

A numeric vector with radian values.

Details

Converting other units of decimal time

convert_rad assumes that any numeric value on x represents decimal hours. Check the instructions on convert_to_date_time details section to learn how to convert other units of decimal time in radians.

Round-off errors

This function is not optimized to deal with round-off errors. If you need a very precise transformation, convert_to_rad may not attend you.

See also

Examples

## ** conversion of date/time objects to radian values **
x <- lubridate::ymd_hms("2020-01-01 05:25:00")
convert_to_rad(x)

#> [1] 1.41808

#> [1] 1.41808 # Expected
convert_to_rad(x, round = TRUE, digits = 4)

#> [1] 1.4181

#> [1] 1.4181 # Expected

x <- c(lubridate::dhours(1), lubridate::dhours(23))
convert_to_rad(x)

#> [1] 0.2617994 6.0213859

#> [1] 0.2617994 6.0213859 # Expected

x <- hms::as_hms("03:45:23")
convert_to_rad(x)

#> [1] 0.9834203

#> [1] 0.9834203 # Expected

## ** conversion of decimal hours to radian value **
x <- 1
convert_to_rad(x)

#> [1] 0.2617994

#> [1] 0.2617994 # Expected
convert_to_rad(x, round = TRUE, digits = 2)

#> [1] 0.26

#> [1] 0.26 # Expected

x <- c(24, 1)
convert_to_rad(x)

#> [1] 6.2831853 0.2617994

#> [1] 6.2831853 0.2617994 # Expected