Compute MCTQ average weekly light exposure

le_week() computes the average weekly light exposure for the standard version of the Munich Chronotype Questionnaire (MCTQ).

le_week(le_w, le_f, wd)

Arguments

le_w
A Duration object corresponding to the light exposure on work days value from a standard version of the MCTQ questionnaire.

le_f
A Duration object corresponding to the light exposure on work-free days value from a standard version of the MCTQ questionnaire.

wd
An integerish integer or numeric object corresponding to the number of work days per week value from a standard or micro version of the MCTQ questionnaire.

Value

A Duration object corresponding to the weighted mean of le_w and le_f with wd and fd(wd) as weights.

Details

Standard MCTQ functions were created following the guidelines in Roenneberg, Wirz-Justice, & Merrow (2003), Roenneberg, Allebrandt, Merrow, & Vetter (2012), and The World Wide Experimental Platform (theWeP, n.d.).

\(\mu\)MCTQ functions were created following the guidelines in Ghotbi et.al (2020), in addition to the guidelines used for the standard MCTQ.

MCTQ\(^{Shift}\) functions were created following the guidelines in Juda, Vetter, & Roenneberg (2013), in addition to the guidelines used for the standard MCTQ.

See References section to learn more.

Class requirements

The mctq package works with a set of object classes specially created to hold time values. This classes can be found in the hms and lubridate packages. If your data do not conform to the object classes required, you can use convert() to convert it.

Rounding and fractional time

Some operations may produce an output with fractional time (e.g. "19538.3828571429s (~5.43 hours)"; 01:15:44.505). If you want, you can round it with round_time().

Our recommendation is to avoid rounding, but, if you do, make sure that you only round your values after all computations are done, that way you can avoid round-off errors.

Guidelines

Roenneberg, Allebrandt, Merrow, & Vetter (2012) and theWeP (n.d.) guidelines for le_week() (\(LE_{week}\)) computation are as follow.

Notes

The average weekly light exposure is the weighted average of the light exposure on work and work-free days in a week.

If you are visualizing this documentation in plain text (ASCII), you may have some trouble to understand the equations. If you want a better viewer, you can see this documentation on the package website.

Computation

$$\frac{LE_W \times WD + LE_F \times FD}{7}$$

Where:

\(LE_W\) = light exposure on workdays.

\(LE_F\) = light exposure on work-free days.

\(WD\) = number of workdays per week ("I have a regular work schedule and work ___ days per week").

\(FD\) = number of work-free days per week.

* \(W\) = workdays; \(F\) = work-free days.

Missing sections in standard and micro MCTQ versions

Although the standard and micro versions of the MCTQ ask for subjects to complete the workdays and work-free days sections, even when he/she do not have a regular work schedule (wd = 0) or have a 7 day/week work schedule (wd = 7), some of them may still end skipping one of those parts of the questionnaire. In those cases, sd_week(), msf_sc(), sloss_week(), sjl(), and le_week() will produce NA (Not Available) as output. That's because their computation combines workdays and work-free days variables.

A NA value is the correct output for the functions above when wd > 0 & wd < 7, but it may not be when wd == 0 or wd == 7. There are different approaches to deal with those cases. See vignette("missing-sections", package = "mctq") to learn more.

References

Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen, D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T. (2020). The \(\mu\)MCTQ: an ultra-short version of the Munich ChronoType Questionnaire. Journal of Biological Rhythms, 35(1), 98-110. doi: 10.1177/0748730419886986 .

Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType Questionnaire for shift-workers (MCTQ\(^{Shift}\)). Journal of Biological Rhythms, 28(2), 130-140. doi: 10.1177/0748730412475041 .

Roenneberg T., Allebrandt K. V., Merrow M., & Vetter C. (2012). Social jetlag and obesity. Current Biology, 22(10), 939-43. doi: 10.1016/j.cub.2012.03.038 .

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks: daily temporal patterns of human chronotypes. Journal of Biological Rhythms, 18(1), 80-90. doi: 10.1177/0748730402239679 .

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from https://www.thewep.org/documentations/mctq/.

See also

Examples

## __ Scalar example __
le_w <- lubridate::dhours(1.5)
le_f <- lubridate::dhours(3.7)
wd <- 5
le_week(le_w, le_f, wd)

#> [1] "7662.85714285714s (~2.13 hours)"

#> [1] "7662.85714285714s (~2.13 hours)" # Expected

le_w <- lubridate::dhours(3)
le_f <- lubridate::dhours(1.5)
wd <- 6
le_week(le_w, le_f, wd)

#> [1] "10028.5714285714s (~2.79 hours)"

#> [1] "10028.5714285714s (~2.79 hours)" # Expected

le_w <- lubridate::dhours(5.6)
le_f <- lubridate::as.duration(NA)
wd <- 3
le_week(le_w, le_f, wd)

#> [1] NA

#> [1] NA # Expected

## __ Vector example __
le_w <- c(lubridate::dhours(3), lubridate::dhours(2.45))
le_f <- c(lubridate::dhours(3), lubridate::dhours(3.75))
wd <- c(4, 5)
le_week(le_w, le_f, wd)

#> [1] "10800s (~3 hours)"               "10157.1428571429s (~2.82 hours)"

#> [1] "10800s (~3 hours)" # Expected
#> [2] "10157.1428571429s (~2.82 hours)" # Expected

## __ Checking second output from vector example __
if (requireNamespace("stats", quietly = TRUE)) {
    i <- 2
    x <- c(le_w[i], le_f[i])
    w <- c(wd[i], fd(wd[i]))
    lubridate::as.duration(stats::weighted.mean(x, w))
#> [1] "10157.1428571429s (~2.82 hours)" # Expected
}

#> [1] "10157.1428571429s (~2.82 hours)"

## __ Converting the output to `hms` __
le_w <- lubridate::dhours(1.25)
le_f <- lubridate::dhours(6.23)
wd <- 3
x <- le_week(le_w, le_f, wd)
x

#> [1] "14744.5714285714s (~4.1 hours)"

#> [1] "14744.5714285714s (~4.1 hours)" # Expected
convert(x, "hms")

#> 04:05:44.571429

#> 04:05:44.571429 # Expected

## __ Rounding the output at the seconds level __
le_w <- lubridate::dhours(3.4094)
le_f <- lubridate::dhours(6.2345)
wd <- 2
x <- le_week(le_w, le_f, wd)
x

#> [1] "19538.3828571429s (~5.43 hours)"

#> [1] "19538.3828571429s (~5.43 hours)" # Expected
round_time(x)

#> [1] "19538s (~5.43 hours)"

#> [1] "19538s (~5.43 hours)" # Expected