Load a data file to R

load_data() is a wrapper function for readr::read_delim() to help simple data loading. It can read any kind of delimited data in addition to .RData files.

If this function doesn't work for your file, we recommend using the readr::readr package to load it. If you're using RStudio IDE, you can also go to Import Dataset, on the environment tab, to load your data.

load_data(
  file = as.character(tcltk::tkgetOpenFile()),
  delim = ",",
  na = c("", " ", "NA"),
  col_types = readr::cols(.default = "c"),
  trim_ws = TRUE,
  skip = 0,
  skip_empty_rows = TRUE
)

Arguments

file
A string indicating the file address to load. If unassigned, a dialog window will open allowing browsing (may not work in some OSs).

delim
(optional) A string containing the field separator in file (default: ",").

na
(optional) A character vector indicating values that must be interpreted as NA (default: c("", " ", "NA")).

col_types
(optional) A NULL value or a readr::cols() specification to set how file columns must be treated. Check readr::read_delim() to learn more (default: readr::cols(.default = "c"), which imports all columns as character).

trim_ws
(optional) A logical value indicating if leading and/or trailing whitespaces must be trimmed from each field (default: TRUE).

skip
(optional) An integerish value, equal or greater than 0, indicating the number of rows to skip when reading file (default: 0).

skip_empty_rows
(optional) A logical value indicating if blank rows must be ignored altogether. If this option is TRUE, then blank rows will not be represented at all. If it is FALSE then they will be represented by NA values in all the columns (default: TRUE).

Value

An invisible tibble with data imported from file.

Details

.csv and .tsv files

Common delimiters for .csv (Comma-Separated Values) are , and ;. For .tsv (Tab-Separated Values) use "\t".

.RData files

load_data() can also read .RData files. In this case, only the file argument is needed, all other arguments are ignored.

If you need to read .rda files, use the base::load() function.

See also

Examples

if (FALSE) {
data <- load_data(raw_data("std_mctq.csv"))
}