A MCTQ core dataset

A fictional MCTQ core (the original) dataset for testing and learning.

This data was created according to Roenneberg, Wirz-Justice & Merrow (2003) and the guidelines of The World Wide Experimental Platform (theWeP, n.d.). See References and Details sections to learn more.

The naming of the variables took into account the standard used by theWeP and the guidelines of the tidyverse style guide.

data(mctq)

Format

A data frame with 10 rows and 18 variables:

regular_work_schedule (logical)

A logical value indicating if the subject have a regular work schedule.Statement (EN): "I have a regular work schedule (this includes being, for example, a housewife or househusband)").

wd (integer)

Number of workdays per week.Statement (EN): "I have a regular work schedule and work ___ days per week").

bt_w (hms)

Local time of going to bed on workdays.Statement (EN): "I go to bed at ___ o'clock'".

s_prep_w (hms)

Local time of preparing to sleep on workdays.Statement (EN): "I actually get ready to fall asleep at ___ o'clock".

s_lat_w (Duration)

Sleep latency on workdays.Statement (EN): "I need ___ minutes to fall asleep".

se_w (hms)

Sleep end on workdays.Statement (EN): "I wake up at ___ o'clock".

si_w (Duration)

Sleep inertia on workdays.Statement (EN): "After ___ minutes, I get up".

alarm_w (logical)

A logical value indicating if the subject use an alarm clock on workdays.Statement (EN): "I use an alarm clock on workdays".

wake_before_alarm_w (logical)

A logical value indicating if the subject regularly wake up BEFORE the alarm rings.Statement (EN): "I regularly wake up BEFORE the alarm rings".

le_w (Duration)

Light exposure on workdays.Statement (EN): "On average, I spend the following amount of time outdoors in daylight (without a roof above my head)".

bt_f (hms)

Local time of going to bed on work-free days.Statement (EN): "I go to bed at ___ o'clock'".

s_prep_f (hms)

Local time of preparing to sleep on work-free daysStatement (EN): "I actually get ready to fall asleep at ___ o'clock".

s_lat_f (Duration)

Sleep latency on work-free days.Statement (EN): "I need ___ minutes to fall asleep".

se_f (hms)

Sleep end on work-free days.Statement (EN): "I wake up at ___ o'clock".

si_f (Duration)

Sleep inertia on work-free days.Statement (EN): "After ___ minutes, I get up".

alarm_f (logical)

A logical value indicating if the subject use an alarm clock on work-free days.Statement (EN): "My wake-up time is due to the use of an alarm clock".

reasons_f (character)

Particular reasons why the subject cannot freely choose his sleep times on work-free days.Statement (EN): "There are particular reasons why I cannot freely choose my sleep times on free days".

le_f (Duration)

Light exposure on work-free days.Statement (EN): "On average, I spend the following amount of time outdoors in daylight (without a roof above my head)".

Source

Prepared by Daniel Vartanian (package's author).

Details

To learn more about the Munich Chronotype Questionnaire (MCTQ), cf. Roenneberg, Wirz-Justice & Merrow (2003).

To know about the different MCTQ versions, cf. Juda, Vetter & Roenneberg (2013) and Ghotbi et.al (2020).

If you curious about the computations and want to have access to the full questionnaire, cf. The Worldwide Experimental Platform (n.d.).

References

Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType Questionnaire for shift-workers (MCTQShift). Journal of Biological Rhythms, 28(2), 130-140. doi: 10.1177/0748730412475041.

Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen, D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T. (2020). The \(\mu\)MCTQ: an ultra-short version of the Munich ChronoType Questionnaire. Journal of Biological Rhythms, 35(1), 98-110. doi: 10.1177/0748730419886986.

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks: daily temporal patterns of human chronotypes. Journal of Biological Rhythms, 18(1), 80-90. doi: 10.1177/0748730402239679.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from https://www.thewep.org/documentations/mctq/.