A fictional \(\mu\)MCTQ dataset

A fictional dataset composed by micro Munich Chronotype Questionnaire (MCTQ) basic/measurable and computed variables for testing and learning purposes.

This data was created following the guidelines in Ghotbi et.al (2020), in addition to the guidelines in Roenneberg, Wirz-Justice, & Merrow (2003), Roenneberg, Allebrandt, Merrow, & Vetter (2012), and The World Wide Experimental Platform (theWeP, n.d.). See References and Details sections to learn more.

micro_mctq

Format

A tibble with 17 columns and 50 rows:

id

A unique integer value for the purpose of identifying each subject in the dataset.

Type: Control.

R class: integer.

shift_work

A logical value indicating if the subject have been a shift- or night-worker in the past three months.

Statement (EN): "I have been a a shift- or night-worker in the past three months".

Type: Basic.

R class: logical.

wd

Number of workdays per week.

Statement (EN): "Normally, I work ___ days/week".

Type: Basic.

R class: integer.

fd

Number of work-free days per week.

Type: Computed.

R class: integer.

so_w

Sleep onset on workdays.

Statement (EN): "On WORKDAYS ... I normally fall asleep at ___ : ___ AM/PM (this is NOT when you get into bed, but rather when you fall asleep)".

Type: Basic.

R class: hms.

se_w

Sleep end on workdays.

Statement (EN): "On WORKDAYS ... I normally wake up at ___ : ___ AM/PM (this is NOT when you get out of bed, but rather when you wake up)".

Type: Basic.

R class: hms.

sd_w

Sleep duration on workdays.

Type: Computed.

R class: Duration.

msw

Mid-sleep on workdays.

Type: Computed.

R class: hms.

so_f

Sleep onset on work-free days when the subject DON'T use an alarm clock.

Statement (EN): "On WORK-FREE DAYS when I DON'T use an alarm clock ... I normally fall asleep at ___ : ___ AM/PM (this is NOT when you get into bed, but rather when you fall asleep)".

Type: Basic.

R class: hms.

se_f

Sleep end on work-free days when the subject DON'T use an alarm clock.

Statement (EN): "On WORK-FREE DAYS when I DON'T use an alarm clock ... I normally wake up at ___ : ___ AM/PM (this is NOT when you get out of bed, but rather when you wake up)".

Type: Basic.

R class: hms.

sd_f

Sleep duration on work-free days when the subject DON'T use an alarm clock.

Type: Computed.

R class: Duration.

msf

Mid-sleep on work-free days when the subject DON'T use an alarm clock.

Type: Computed.

R class: hms.

sd_week

Average weekly sleep duration.

Type: Computed.

R class: Duration.

msf_sc

Chronotype or corrected mid-sleep on work-free days.

Type: Computed.

R class: hms.

sloss_week

Weekly sleep loss.

Type: Computed.

R class: Duration.

sjl_rel

Relative social jetlag.

Type: Computed.

R class: Duration.

sjl

Absolute social jetlag.

Type: Computed.

R class: Duration.

Source

Prepared by Daniel Vartanian (package's author).

Details

micro_mctq is a tidied, validated, and transformed version of raw_data("micro_mctq.csv").

Guidelines

To learn more about the Munich Chronotype Questionnaire (MCTQ), cf. Roenneberg, Wirz-Justice, & Merrow (2003), Roenneberg, Allebrandt, Merrow, & Vetter (2012), Roenneberg et al. (2015), and Roenneberg, Pilz, Zerbini, & Winnebeck (2019).

To know about different MCTQ versions, cf. Juda, Vetter, & Roenneberg (2013) and Ghotbi et.al (2020).

If you are curious about the variable computations and want to have access to the full questionnaire, cf. The Worldwide Experimental Platform (n.d.).

Data building and data wrangling

This dataset was created by randomized sampling (see random_mctq()) and by manual insertions of especial cases. Its purpose is to demonstrate common cases and data issues that researchers may find in their MCTQ data, in addition to be a suggested data structure for MCTQ data.

You can see the micro_mctq build and data wrangling processes at https://github.com/gipsousp/mctq/blob/master/data-raw/micro_mctq.R.

Variable naming

The naming of the variables took into account the naming schemes used in MCTQ articles, in addition to the guidelines of the tidyverse style guide.

Variable classes

The mctq package works with a set of object classes specially created to hold time values. This classes can be found in the hms and lubridate packages.

References

Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen, D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T. (2020). The \(\mu\)MCTQ: an ultra-short version of the Munich ChronoType Questionnaire. Journal of Biological Rhythms, 35(1), 98-110. doi: 10.1177/0748730419886986 .

Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType Questionnaire for shift-workers (MCTQ\(^{Shift}\)). Journal of Biological Rhythms, 28(2), 130-140. doi: 10.1177/0748730412475041 .

Roenneberg T., Allebrandt K. V., Merrow M., & Vetter C. (2012). Social jetlag and obesity. Current Biology, 22(10), 939-43. doi: 10.1016/j.cub.2012.03.038 .

Roenneberg, T., Keller, L. K., Fischer, D., Matera, J. L., Vetter, C., & Winnebeck, E. C. (2015). Human activity and rest in situ. In A. Sehgal (Ed.), Methods in enzymology (Vol. 552, pp. 257-283). London, UK: Academic Press. doi: 10.1016/bs.mie.2014.11.028 .

Roenneberg, T., Pilz, L. K., Zerbini, G., & Winnebeck, E. C. (2019). Chronotype and social jetlag: a (self-) critical review. Biology, 8(3), 54. doi: 10.3390/biology8030054 .

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks: daily temporal patterns of human chronotypes. Journal of Biological Rhythms, 18(1), 80-90. doi: 10.1177/0748730402239679 .

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from https://www.thewep.org/documentations/mctq/.

See also