Round time values

round_time() takes a Duration, Period, or hms object and round it at the seconds level.

round_time(x)

Arguments

x
An object belonging to one of the following classes: Duration, Period, or hms.

Value

A time object with the time rounded at the seconds level.

Details

round_time() convert time to seconds and them uses base::round() to round it. That is to say that round_time() uses the same IEC 60559 standard ("go to the even digit") for rounding off a 5. Therefore round(0.5) is 0 and round(-1.5) is -2. See ?round to learn more.

See also

Examples

## __ Scalar example __
lubridate::dhours(1.45678696454)

#> [1] "5244.433072344s (~1.46 hours)"

#> [1] "5244.433072344s (~1.46 hours)" # Expected
round_time(lubridate::dhours(1.45678696454))

#> [1] "5244s (~1.46 hours)"

#> [1] "5244s (~1.46 hours)" # Expected

lubridate::microseconds(2454876956)

#> [1] "2454.876956S"

#> [1] "2454.876956S" # Expected
hms::as_hms(as.numeric(lubridate::microseconds(2454876956)))

#> 00:40:54.876956

#> 00:40:54.876956 # Expected
round_time(lubridate::microseconds(2454876956))

#> [1] "40M 55S"

#> [1] "40M 55S" # Expected

hms::as_hms(12345.6789)

#> 03:25:45.6789

#> 03:25:45.6789 # Expected
round_time(hms::as_hms(12345.6789))

#> 03:25:46

#> 03:25:46 # Expected

## __ Vector example __
x <- c(lubridate::dhours(5.6987), lubridate::dhours(2.6875154))
x

#> [1] "20515.32s (~5.7 hours)"    "9675.05544s (~2.69 hours)"

#> [1] "20515.32s (~5.7 hours)"    "9675.05544s (~2.69 hours)" # Expected
round_time(x)

#> [1] "20515s (~5.7 hours)" "9675s (~2.69 hours)"

#> [1] "20515s (~5.7 hours)" "9675s (~2.69 hours)" # Expected