Find the shortest interval between two hours

shortest_interval() finds and return the shortest interval between two hms or POSIXt objects hours. This is useful for time arithmetic, because there's always two possible intervals between two hour values with no date reference.

longer_interval() do the inverse of shortest_interval(), i.e finds the longer interval between two hours. It's just a wrapper for shortest_interval(x, y, class, inverse = TRUE).

shortest_interval(x, y, class = "hms", inverse = FALSE, quiet = FALSE)

longer_interval(x, y, class = "hms", quiet = FALSE)

Arguments

x, y
A hms or POSIXt object.

class
(optional) a string indicating the object class of the output (default: "hms").

inverse
(optional) a logical value indicating if the function must return a inverse output, i.e the longer interval between x and y.

quiet
(optional) a logical value indicating if warnings or messages must be suppressed (default: FALSE).

Value

A R object, of class indicated on class, with the shortest or longer interval (if inverse = TRUE or longer_interval() is used) between x and y.

Details

The two intervals problem

Given two hours, x and y, in a two day timeline, without date references, there will be always two possible intervals between them, as illustrated below.

To figure out what interval is the shortest or the longer, shortest_interval() checks two scenarios: 1. When x comes before y; and 2. when x comes after y. This only works if x value is smaller than y, therefore, the function will make sure to swap x and y values if the latter assumption is not true.

Because shortest_interval() objective is to find the shortest interval, if x and y are equal, the shortest interval will have a length of 0 hours, resulting in a interval from x to x. But, if inverse = TRUE or longer_interval() is use instead, the latter condition will return a interval with 24 hours of length (from x to x + 1 day).

In cases when x and y distance themselves by 12 hours, there will be no shortest or longer interval (they are equal). In this cases, shortest_interval() and longer_interval() will return the same value.

             day 1                        day 2
     x                  y         x                  y
   06:00              22:00     06:00              22:00
-----|------------------|---------|------------------|----->
              16h           8h             16h
          longer int.   shorter int.    longer int.

              day 1                      day 2
     y                   x       y                   x
   13:00               08:00   13:00               08:00
-----|-------------------|-------|-------------------|----->
              19h           5h            19h
          longer int.   shorter int.   longer int.

    x,y             x,y             x,y             x,y
     x               y               x               y
   10:00           10:00           10:00           10:00
-----|---------------|---------------|---------------|----->
    0h              0h              0h              0h
            24h             24h             24h

              day 1                      day 2
     y               x               y               x
   12:00           00:00           12:00           00:00
-----|---------------|---------------|---------------|----->
            12h             12h             12h

class argument

shortest_interval() is integrated with convert(). That way you can choose what class of object your prefer as output.

Valid class values are: "Duration", "Period", "difftime", "hms", and "Interval" (case insensitive).

POSIXt objects

POSIXt values passed as argument to x or y will be strip of their dates. Only the hours will be considered.

NA values

shortest_interval() will return NA if x or y are NA.

See also

Examples

## __ Scalar example __
x <- hms::parse_hm("23:00")
y <- hms::parse_hm("01:00")
shortest_interval(x, y)

#> 02:00:00

#> 02:00:00 # Expected

x <- lubridate::as_datetime("1985-01-15 12:00:00")
y <- lubridate::as_datetime("2020-09-10 12:00:00")
shortest_interval(x, y)

#> 00:00:00

#> 00:00:00 # Expected

## __ Vector example __
x <- c(hms::parse_hm("15:30"), hms::parse_hm("21:30"))
y <- c(hms::parse_hm("19:30"), hms::parse_hm("04:00"))
shortest_interval(x, y)

#> 04:00:00
#> 06:30:00

#> 04:00:00 # Expected
#> 06:30:00 # Expected

## __ Finding the longer interval between two hours __
x <- lubridate::parse_date_time("01:10:00", "HMS")
y <- lubridate::parse_date_time("11:45:00", "HMS")
shortest_interval(x, y, inverse = TRUE)

#> 13:25:00

#> 13:25:00 # Expected

x <- lubridate::as_datetime("1915-02-14 05:00:00")
y <- lubridate::as_datetime("1970-07-01 05:00:00")
longer_interval(x, y)

#> 24:00:00

#> 24:00:00 # Expected

## __ Changing the output object class __
x <- as.POSIXct("1988-10-05 02:00:00")
y <- as.POSIXlt("2100-05-07 13:30:00")
shortest_interval(x, y, "Interval")

#> [1] 1970-01-01 02:00:00 UTC--1970-01-01 13:30:00 UTC

#> [1] 1970-01-01 02:00:00 UTC--1970-01-01 13:30:00 UTC # Expected
longer_interval(x, y, "Duration")

#> [1] "45000s (~12.5 hours)"

#> [1] "45000s (~12.5 hours)" # Expected
shortest_interval(x, y, "Period")

#> [1] "11H 30M 0S"

#> [1] "11H 30M 0S" # Expected
longer_interval(x, y, "hms")

#> 12:30:00

#> 12:30:00" # Expected