A fictional standard MCTQ dataset

A fictional dataset, for testing and learning purposes, composed of basic/measurable and computed variables of the Munich ChronoType Questionnaire (MCTQ) standard version.

This data was created following the guidelines in Roenneberg, Wirz-Justice, & Merrow (2003), Roenneberg, Allebrandt, Merrow, & Vetter (2012), Jankowski (2017), and The Worldwide Experimental Platform (n.d.). See the References and Details sections to learn more.

Usage

std_mctq

Format

A tibble with 39 columns and 50 rows:

id

A unique integer value to identify each respondent in the dataset.

Type: Control.

R class: integer.

work

A logical value indicating if the respondent has a regular work schedule.

Statement (EN): "I have a regular work schedule (this includes being, for example, a housewife or househusband): Yes ( ___ ) No ( ___ )".

Type: Basic.

R class: logical.

wd

Number of workdays per week.

Statement (EN): "I have a regular work schedule and work ___ days per week".

Type: Basic.

R class: integer.

fd

Number of work-free days per week.

Type: Computed.

R class: integer.

bt_w

Local time of going to bed on workdays.

Statement (EN): "I go to bed at ___ o'clock'".

Type: Basic.

R class: hms.

sprep_w

Local time of preparing to sleep on workdays.

Statement (EN): "I actually get ready to fall asleep at ___ o'clock".

Type: Basic.

R class: hms.

slat_w

Sleep latency or time to fall asleep after preparing to sleep on workdays.

Statement (EN): "I need ___ minutes to fall asleep".

Type: Basic.

R class: Duration.

so_w

Local time of sleep onset on workdays.

Type: Computed.

R class: hms.

se_w

Local time of sleep end on workdays.

Statement (EN): "I wake up at ___ o'clock".

Type: Basic.

R class: hms.

si_w

"Sleep inertia" on workdays.

Despite the name, this variable represents the time the respondent takes to get up after sleep end.

Statement (EN): "After ___ minutes, I get up".

Type: Basic.

R class: Duration.

gu_w

Local time of getting out of bed on workdays.

Type: Computed.

R class: hms.

alarm_w

A logical value indicating if the respondent uses an alarm clock to wake up on workdays.

Statement (EN): "I use an alarm clock on workdays: Yes ( ___ ) No ( ___ )".

Type: Basic.

R class: logical.

wake_before_w

A logical value indicating if the respondent regularly wakes up before the alarm rings on workdays.

Statement (EN): "If "Yes": I regularly wake up BEFORE the alarm rings: Yes ( ___ ) No ( ___ )".

Type: Basic.

R class: logical.

sd_w

Sleep duration on workdays.

Type: Computed.

R class: Duration.

tbt_w

Total time in bed on workdays.

Type: Computed.

R class: Duration.

le_w

Light exposure on workdays.

Statement (EN): "On average, I spend the following amount of time outdoors in daylight (without a roof above my head)".

Type: Extra.

R class: Duration.

msw

Local time of mid-sleep on workdays.

Type: Computed.

R class: hms.

bt_f

Local time of going to bed on work-free days.

Statement (EN): "I go to bed at ___ o'clock'".

Type: Basic.

R class: hms.

sprep_f

Local time of preparing to sleep on work-free days.

Statement (EN): "I actually get ready to fall asleep at ___ o'clock".

Type: Basic.

R class: hms.

slat_f

Sleep latency or time to fall asleep after preparing to sleep on work-free days.

Statement (EN): "I need ___ minutes to fall asleep".

Type: Basic.

R class: Duration.

so_f

Local time of sleep onset on work-free days.

Type: Computed.

R class: hms.

se_f

Local time of sleep end on work-free days.

Statement (EN): "I wake up at ___ o'clock".

Type: Basic.

R class: hms.

si_f

"Sleep inertia" on work-free days.

Despite the name, this variable represents the time the respondent takes to get up after sleep end.

Statement (EN): "After ___ minutes, I get up".

Type: Basic.

R class: Duration.

gu_f

Local time of getting out of bed on work-free days.

Type: Computed.

R class: hms.

alarm_f

A logical value indicating if the respondent uses an alarm clock to wake up on work-free days.

Statement (EN): "My wake-up time is due to the use of an alarm clock: Yes ( ___ ) No ( ___ )".

Type: Basic.

R class: logical.

reasons_f

A logical value indicating if the respondent has any particular reasons for why they cannot freely choose their sleep times on work-free days.

Statement (EN): "There are particular reasons why I cannot freely choose my sleep times on free days: Yes ( ___ ) No ( ___ )".

Type: Basic.

R class: logical.

reasons_why_f

Particular reasons for why the respondent cannot freely choose their sleep times on work-free days.

Statement (EN): "If "Yes": Child(ren)/pet(s) ( ___ ) Hobbies ( ___ ) Others ( ___ ), for example: ___".

Type: Basic.

R class: character.

sd_f

Sleep duration on work-free days.

Type: Computed.

R class: Duration.

tbt_f

Total time in bed on work-free days.

Type: Computed.

R class: Duration.

le_f

Light exposure on work-free days.

Statement (EN): "On average, I spend the following amount of time outdoors in daylight (without a roof above my head)".

Type: Extra.

R class: Duration.

msf

Local time of mid-sleep on work-free days.

Type: Computed.

R class: hms.

sd_week

Average weekly sleep duration.

Type: Computed.

R class: Duration.

sloss_week

Weekly sleep loss.

Type: Computed.

R class: Duration.

le_week

Average weekly light exposure.

Type: Computed.

R class: Duration.

msf_sc

Sleep-corrected local time of mid-sleep on work-free days.

Type: Computed.

R class: hms.

sjl_rel

Relative social jetlag.

Type: Computed.

R class: Duration.

sjl

Absolute social jetlag.

Type: Computed.

R class: Duration.

sjl_sc_rel

Jankowski's relative sleep-corrected social jetlag.

Type: Computed.

R class: Duration.

sjl_sc

Jankowski's sleep-corrected social jetlag.

Type: Computed.

R class: Duration.

Source

Created by Daniel Vartanian (package author).

Details

std_mctq is a tidied, validated, and transformed version of raw_data("std_mctq.csv").

Guidelines

To learn more about the Munich ChronoType Questionnaire (MCTQ), see Roenneberg, Wirz-Justice, & Merrow (2003), Roenneberg, Allebrandt, Merrow, & Vetter (2012), Roenneberg et al. (2015), and Roenneberg, Pilz, Zerbini, & Winnebeck (2019).

To know about different MCTQ versions, see Juda, Vetter, & Roenneberg (2013) and Ghotbi et al. (2020).

To learn about the sleep-corrected social jetlag, see Jankowski (2017).

If you're curious about the variable computations and want to have access to the full questionnaire, see The Worldwide Experimental Platform (n.d.).

Data building and data wrangling

This dataset was created by randomized sampling (see random_mctq()) and by manual insertions of special cases. Its purpose is to demonstrate common cases and data issues that researchers may find in their MCTQ data, in addition to be a suggested data structure for MCTQ data.

You can see the std_mctq build and data wrangling processes here.

Variable naming

The naming of the variables took into account the naming scheme used in MCTQ publications, in addition to the guidelines of the tidyverse style guide.

Variable classes

The mctq package works with a set of object classes specially created to hold time values. These classes can be found in the hms and lubridate package.

Duration objects

If you prefer to view Duration objects as hms objects, run pretty_mctq(std_mctq).

References

Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen, D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T. (2020). The \(\mu\)MCTQ: an ultra-short version of the Munich ChronoType Questionnaire. Journal of Biological Rhythms, 35(1), 98-110. doi:10.1177/0748730419886986

Jankowski K. S. (2017). Social jet lag: sleep-corrected formula. Chronobiology International, 34(4), 531-535. doi:10.1080/07420528.2017.1299162

Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType Questionnaire for shift-workers (MCTQ\(^{Shift}\)). Journal of Biological Rhythms, 28(2), 130-140. doi:10.1177/0748730412475041

Roenneberg T., Allebrandt K. V., Merrow M., & Vetter C. (2012). Social jetlag and obesity. Current Biology, 22(10), 939-43. doi:10.1016/j.cub.2012.03.038

Roenneberg, T., Keller, L. K., Fischer, D., Matera, J. L., Vetter, C., & Winnebeck, E. C. (2015). Human activity and rest in situ. In A. Sehgal (Ed.), Methods in Enzymology (Vol. 552, pp. 257-283). Academic Press. doi:10.1016/bs.mie.2014.11.028

Roenneberg, T., Pilz, L. K., Zerbini, G., & Winnebeck, E. C. (2019). Chronotype and social jetlag: a (self-) critical review. Biology, 8(3), 54. doi:10.3390/biology8030054

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks: daily temporal patterns of human chronotypes. Journal of Biological Rhythms, 18(1), 80-90. doi:10.1177/0748730402239679

The Worldwide Experimental Platform (n.d.). MCTQ. https://www.thewep.org/documentations/mctq/

See also