Compute MCTQ total time in bed

tbt() computes the total time in bed for the standard Munich Chronotype Questionnaire (MCTQ).

tbt(bt, gu)

Arguments

bt
A hms vector corresponding to the local time of going to bed of a standard MCTQ questionnaire.

gu
A hms vector corresponding to the local time of getting out of bed of a standard MCTQ questionnaire (you can use gu() to compute it).

Value

A hms vector corresponding to the difference between gu and bt rolled on a 24 hours clock basis.

Details

This function was created according to Roenneberg, Wirz-Justice & Merrow (2003) and guidelines of The World Wide Experimental Platform (theWeP, n.d.).

Class requirements

The mctq package works with a set of object classes specially created to hold time values. This classes can be found on hms::hms-package and lubridate::lubridate-package. If your data do not conform to the object classes required, you can use convert_to() to convert it.

Rounding and fractional time

Some operations may produce an output with fractional time (e.g. "19538.3828571429s (~5.43 hours)"; 01:15:44.505). If you want, you can round it with round_time().

We recommend rounding values only after all computations are done, that way you can avoid round-off errors.

Please keep in mind that this package don't assume as true a level of accuracy beyond the seconds level. That is not needed for computing MCTQ variables.

References

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks: daily temporal patterns of human chronotypes. Journal of Biological Rhythms, 18(1), 80-90. doi: 10.1177/0748730402239679.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from https://www.thewep.org/documentations/mctq/.

See also

Examples

## __ Scalar example __
tbt(hms::parse_hms("22:10:00"), hms::parse_hms("06:15:00"))

#> [1] "29100s (~8.08 hours)"

#> 08:05:00 # Expected
tbt(hms::parse_hms("01:20:00"), hms::parse_hms("14:00:00"))

#> [1] "45600s (~12.67 hours)"

#> 12:40:00 # Expected
tbt(hms::as_hms(NA), hms::parse_hms("07:20:00"))

#> [1] NA

#> NA # Expected

## __ Vectorized example __
bt <- c(hms::parse_hms("23:50:00"), hms::parse_hms("02:30:00"))
gu <- c(hms::parse_hms("09:30:00"), hms::parse_hms("11:25:00"))
tbt(bt, gu)

#> [1] "34800s (~9.67 hours)" "32100s (~8.92 hours)"

#> [1] 09:40:00 # Expected
#> [1] 08:55:00 # Expected