Tidy build_std_mctq() output

tidy_std_mctq tidy the output of build_std_mctq().

tidy_std_mctq(write = FALSE)

Arguments

write
A logical value indicating if tidy_std_mctq must write a std_mctq.rda file to "./data/" (optional) (default: FALSE).

Value

An invisible tibble.

Details

Here the process of tiding a dataset is understood as transforming it in input data, like described in Loo and Jonge (2018.

Please note that input data is not the same as valid data. To get a valid std_mctq data, run validate_std_mctq().

To learn more about the concept of tidy data, c.f. Wickham (2014 and Wickham and Grolemund (n.d..

References

Van der Loo, M., & De Jonge, E. (2018). Statistical data cleaning with applications in R. Hooboken, NJ: John Wiley & Sons. doi: 10.1002/9781118897126.

Wickham, H. (2014). Tidy Data. Journal of Statistical Software, 59(10), 1-23. doi: 10.18637/jss.v059.i10.

Wickham, H, & Grolemund. (n.d.). R for data science. Sebastopol, CA: O'Reilly Media. https://r4ds.had.co.nz.

See also

Examples

if (FALSE) {
utils::View(tidy_std_mctq())
tidy_std_mctq(write = TRUE)
}