Write a delimited or a .RData file

write_data() is a wrapper function for readr::write_delim() to help with simple data exporting. It can write any kind of delimited data in addition to .RData files.

write_data() also deals with some common annoyances, like the need to mutate objects of class Duration or Period to a more convenient format (hms) when exporting data to text.

If this function doesn't work for your data, we recommend using the readr::readr package to export it.

write_data(
  data,
  file = as.character(tcltk::tkgetSaveFile()),
  type = "csv",
  delim = ",",
  na = "NA",
  append = FALSE,
  col_names = !append,
  decimal = "."
)

Arguments

data
A data frame.

file
A string indicating the file address where to write data. If unassigned, a dialog window will open allowing browsing (may not work in some OSs).

type
(optional) A csv, tsv or RData string indicating the type of file to write (default: "csv").

delim
(optional) A string indicating the field separator for writing delimiter-separated files (default: ",").

na
(optional) A string indicating the value for NAs. Missing values will never be quoted; strings with the same value as NA will always be quoted (default: "NA").

append
(optional) A logical value indicating if the function must append data in file or overwrite it. If FALSE, will overwrite existing file. If TRUE, will append to the existing file. In both cases, if the file does not exist a new file is created (default: FALSE).

col_names
(optional) A logical value indicating if the column names must be placed at the top of the file. If not specified, col_names will take the opposite value given to append (default: !append).

decimal
(optional) A "." or "," (European format) string, indicating a new decimal separator when writing delimiter-separated values files (default: ".").

Value

Write data in an address indicated in file. No value is returned.

Details

type argument

The type argument is only needed if a .csv, .tsv or .RData file extension is not found in file. If one of this extensions is present, write_data() will change the type value for the appropriated format.

.csv and .tsv files

Common delimiters for .csv (Comma-Separated Values) are , and ;. For .tsv (Tab-Separated Values) use "\t".

Note that if type == "tsv", write_data() will invariably change the delimiter for "\t".

.RData files

write_data() can also write .RData files. In this case, the file will contain the data in data with no changes.

If you need to write .rda files, use the base::save() function.

See also

Examples

if (FALSE) write_data(std_mctq)