A \(\mu\)MCTQ dataset

UNDER DEVELOPMENT

A fictional \(\mu\)MCTQ dataset for testing and learning.

This data was created according to Ghotbi et.al (2020). See References and Details sections to learn more.

data(micro_mctq)

Format

A data frame with ? rows and ? variables:

variable

Description.

Source

Prepared by Daniel Vartanian (package's author).

Details

To learn more about the Munich Chronotype Questionnaire (MCTQ), cf. Roenneberg, Wirz-Justice & Merrow (2003).

To know about the different MCTQ versions, cf. Juda, Vetter & Roenneberg (2013) and Ghotbi et.al (2020).

If you curious about the computations and want to have access to the full questionnaire, cf. The Worldwide Experimental Platform (n.d.).

References

Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType Questionnaire for shift-workers (MCTQShift). Journal of Biological Rhythms, 28(2), 130-140. doi: 10.1177/0748730412475041.

Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen, D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T. (2020). The \(\mu\)MCTQ: an ultra-short version of the Munich ChronoType Questionnaire. Journal of Biological Rhythms, 35(1), 98-110. doi: 10.1177/0748730419886986.

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks: daily temporal patterns of human chronotypes. Journal of Biological Rhythms, 18(1), 80-90. doi: 10.1177/0748730402239679.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from https://www.thewep.org/documentations/mctq/.