Compute MCTQ chronotype/corrected midsleep on free days

msf_sc() computes the chronotype/corrected midsleep on free days for the standard Munich Chronotype Questionnaire (MCTQ).

chronotype() is just a wrapper for msf_sc()

msf_sc(msf, sd_w, sd_f, sd_week, alarm_f)

chronotype(msf, sd_w, sd_f, sd_week, alarm_f)

Arguments

msf
A hms vector corresponding to the mid-sleep on work-free days of a standard MCTQ questionnaire (you can use ms() to compute it).

sd_w
A Duration vector corresponding to the sleep duration on work days of a standard MCTQ questionnaire (you can use sd() to compute it).

sd_f
A Duration vector corresponding to the sleep duration on work-free days of a standard MCTQ questionnaire (you can use sd() to compute it).

sd_week
A Duration vector corresponding to the average weekly sleep duration of a standard MCTQ questionnaire (you can use sd_week() to compute it).

alarm_f
A logical vector corresponding to the alarm clock use on work-free days of a standard MCTQ questionnaire.

Value

A hms vector corresponding to the chronotype/corrected midsleep on free days.

Details

This function was created according to Roenneberg, Wirz-Justice & Merrow (2003) and guidelines of The World Wide Experimental Platform (theWeP, n.d.).

Class requirements

The mctq package works with a set of object classes specially created to hold time values. This classes can be found on hms::hms-package and lubridate::lubridate-package. If your data do not conform to the object classes required, you can use convert_to() to convert it.

References

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks: daily temporal patterns of human chronotypes. Journal of Biological Rhythms, 18(1), 80-90. doi: 10.1177/0748730402239679.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from https://www.thewep.org/documentations/mctq/.

See also

Examples

msf_sc(hms::parse_hms("04:00:00"), lubridate::dhours(6),
       lubridate::dhours(7), lubridate::dhours(6.29), FALSE)

#> Note: method with signature ‘Duration#ANY’ chosen for function ‘-’,
#>  target signature ‘Duration#Duration’.
#>  "ANY#Duration" would also be valid

#> 03:38:42

#> 03:38:42 # Expected
msf_sc(hms::parse_hms("06:30:00"), lubridate::dhours(7.5),
       lubridate::dhours(7.5), lubridate::dhours(7.5), FALSE)

#> 06:30:00

#> 06:30:00 # Expected
msf_sc(hms::parse_hm("01:00:00"), lubridate::dhours(5.5),
       lubridate::dhours(9), lubridate::dhours(6.75), FALSE)

#> 23:52:30

#> 23:52:30 # Expected
msf_sc(hms::parse_hms("05:40:00"), lubridate::dhours(7.5),
       lubridate::dhours(10), lubridate::dhours(8.5), TRUE)

#> NA

#> NA # Expected
x <- c(hms::parse_hms("03:45:00"), hms::parse_hm("04:45:00"))
y <- c(lubridate::dhours(5), lubridate::dhours(6.45))
v <- c(lubridate::dhours(9), lubridate::dhours(10))
w <- c(lubridate::dhours(8.5), lubridate::dhours(9.2))
z <- c(TRUE, TRUE)
msf_sc(x, y, v, w, z)

#> 03:30:00
#> 04:21:00

#> 03:30:00 # Expected
#> 04:21:00 # Expected
chronotype(hms::parse_hms("07:00:00"), lubridate::dhours(6),
           lubridate::dhours(12), lubridate::dhours(9.45), FALSE)

#> 05:43:30

#> 05:43:30 # Expected