Make a MCTQ dataset more presentable

pretty_mctq() helps you to transform your MCTQ data in many ways. See parameters and Details section to learn more.

pretty_mctq(data, round = TRUE, hms = TRUE)

Arguments

data
A data frame.

round
(optional) a logical value indicating if Duration, Period, and hms objects must be rounded at the seconds level (default: TRUE).

hms
(optional) a logical value indicating if Duration, Period, and difftime objects must be converted to hms (default: TRUE).

Value

A transformed data frame, as indicated in arguments.

Details

Rounding

Please note that by rounding MCTQ values you discard data. That is to say that if you need to redo a computation, or do new ones, your values can be off by a couple of seconds (see round-off error).

Round your values only if and when you want to present it in a more clear way, like in graphical representations. You can also round values to facilitate data exporting to text formats (like .csv), but note that this will come with a precision cost.

Note also that pretty_mctq() uses round_time() for rounding. round_time() is based on base::round(), which uses the IEC 60559 standard. For more information see ?round_time.

See also

Examples

data <- data.frame(
    a = 1,
    b = lubridate::duration(1.12345),
    c = lubridate::period(1.12345),
    d = hms::hms(1.12345))

## __ Rounding time objects from `data` __
pretty_mctq(data, round = TRUE, hms = FALSE)

#>   a  b  c        d
#> 1 1 1s 1S 00:00:01

## __ Converting non-`hms` time objects from `data` to `hms` __
pretty_mctq(data, round = FALSE, hms = TRUE)

#>   a              b              c              d
#> 1 1 00:00:01.12345 00:00:01.12345 00:00:01.12345