Walk through distribution plots

qplot_walk() helps you to visually assess the distribution of your data. It uses ggplot2::qplot() to walk through each selected variable from a data frame.

This function requires the grDevices and ggplot2 packages and can only run in interactive mode.

qplot_walk(
  data,
  ...,
  cols = NULL,
  pattern = NULL,
  ignore = "character",
  remove_id = TRUE,
  midday_change = TRUE
)

Arguments

data
An atomic or data frame object.

...
(optional) additional arguments to be passed to ggplot2::qplot().

cols
(optional) (only for data frames) a character vector indicating columns names in data for plotting. If NULL, qplot_walk() will use all columns in data. This setting only works if pattern = NULL (default: NULL).

pattern
(optional) (only for data frames) a string with a regular expression to select columns names in data for plotting. This setting only works if cols = NULL (default: NULL).

ignore
(optional) (only for data frames) a character vector indicating which object classes the function must ignore. This setting can be used with cols and pattern. Assign NULL to disable this behavior (default: "character").

remove_id
(optional) (only for data frames) a logical value indicating if the function must ignore column names in data that match with the regular expression "^id$|[\\._-]id$" (default: TRUE).

midday_change
(optional) a logical value indicating if the function must apply a midday change for hms variables with values greater than 22:00:00 (see Details section to learn more) (default: TRUE).

Details

Requirements

This function requires the grDevices and ggplot2 packages and can only run in interactive mode. The grDevices comes with a standard R installation and is typically loaded by default. Most people also run R interactively.

If you don't have any of the two packages mentioned above, you can install them with install.packages("grDevices") and install.packages("ggplot2").

Plot recover

qplot_walk() erases all plots after it runs. For that reason, the function first emits a dialog message warning the user of this behavior before it runs. If you want to recover a single distribution plot, assign the variable vector on the data argument.

Additional arguments to ggplot2::qplot()

qplot_walk() uses ggplot2 ggplot2::qplot() to generate plots. If you are familiar with ggplot2::qplot(), you can pass additional arguments to the function using the ellipsis argument (...).

Note that x, y and data arguments are reserved for qplot_walk().

Duration, Period, and difftime objects

In order to help with the visualization, qplot_walk() automatically converts Duration, Period, and difftime objects to hms.

Midday change

hms variables with values greater than 22:00:00 will automatically be converted to POSIXct with a midday change, i.e. all values with more than 12 hours will represent day one, and all the rest will represent day 2. This is made in order to better represent values that cross the midnight hour.

You can disable this behavior using midday_change = FALSE.

id variables

qplot_walk() will ignore any variable with the follow name pattern "^id$|[\\._-]id$", i.e any variable named id or that ends with .id, _id, or -id.

You can disable this behavior using remove_id = FALSE.

See also

Examples

if (FALSE) {
qplot_walk(mctq::std_mctq$bt_w)
qplot_walk(mctq::std_mctq)
qplot_walk(mctq::std_mctq, cols = c("bt_w", "msf_sc"))
qplot_walk(mctq::std_mctq, pattern = "_w$")
qplot_walk(datasets::iris)
qplot_walk(datasets::mtcars)
}