Build a random MCTQ case

random_mctq builds a fictional Munich Chronotype Questionnaire (MCTQ) case composed by MCTQ basic/measurable variables.

random_mctq(model = "standard", quiet = FALSE)

Arguments

model
A string indicating the data model to return. Valid values are: "standard", "shift", and "micro" (default: "standard").

quiet
(optional) a logical value indicating if warnings or messages must be suppressed (default: FALSE).

Value

A named list with elements representing each MCTQ basic/measurable variable of the model indicated in model.

Details

The case structure (variable names and classes) are the same as the datasets provided by the mctq package. See std_mctq, micro_mctq and mctq::shift_mctq to learn more.

Requirements

This function requires the stats package. This wont be a issue for most people, since the package comes with a standard R installation.

If you don't have the stats package, you can install it with install.packages("stats").

Cases

Random standard and micro MCTQ cases were created with the general population in mind. The data was set to fit the distributions shown in Roenneberg, Wirz-Justice, & Merrow (2003).

MCTQ\(^{Shift}\) random cases were created based on the shift configuration from "Study Site 1" shown in Vetter, Juda, & Roenneberg (2012) (it's also present in Juda, Vetter, & Roenneberg (2013) supplemental materials).

MCTQ\(^{Shift}\) can be applied to the general population, considering a survey with conditional branching made for shift-workers respondents, but it would be difficult to analyze or even to interpret the data in this case, due to peculiarities of this version of the questionnaire. Given that a fictitious dataset would not be of much use if it did not come close to a real world application, we decided to stick with a configuration that could represent the full potential of the scale.

Standard and micro MCTQ distribution parameters

Based on the general sample distributions shown in Roenneberg, Wirz-Justice, & Merrow (2003) (Table 1, Figure 2, and Figure 7) and assuming normal distributions where the minimal and maximum values are, respectively, \(-3 s\) and \(+3 s\) from the mean (where \(s\) is the standard deviation of the sample), the distribution values for workdays and work-free days for random_mctq(model = "standard") and random_mctq(model = "micro")are as follow.

Notes

Please note that this is just a rough approximation, it by no means represents the same distributions from the base article.

The distribution parameters from other variables not shown here (like sleep latency) were created based on the experience of the package authors with MCTQ data.

This values are just for reference while building the random cases. If you group several random MCTQ cases, this numbers can have some variation.

If you are visualizing this documentation in plain text (ASCII), you may have some trouble to understand the equations. If you want a better viewer, you can see this documentation on the package website.

Workdays

Mid-sleep (\(MSW\))

\(\overline{X}\) \(= \textrm{03:10:00}\) (extracted from the base article)

\(s\) \(= \textrm{00:50:00}\) (extracted from the base article)

\(\textrm{Min}\) \(= \overline{X}_{MSW} - (3 \times s_{MSW}) = \textrm{00:40:00}\)

\(\textrm{Max}\) \(= \overline{X}_{MSW} + (3 \times s_{MSW}) = \textrm{05:40:00}\)

Sleep duration (\(SD\))

\(\overline{X}\) \(= \textrm{07:22:00}\) (extracted from the base article)

\(s\) \(= \textrm{01:09:00}\) (extracted from the base article)

\(\textrm{Min}\) \(= \overline{X}_{SD} - (3 \times s_{SD}) = \textrm{03:53:00}\)

\(\textrm{Max}\) \(= \overline{X}_{SD} + (3 \times s_{SD}) = \textrm{10:47:00}\)

Sleep onset (\(SO\))

\(\overline{X}\) \(= \overline{X}_{MSW} - \frac{\overline{X}_{SD}}{2} = \textrm{23:29:00}\)

\(s\) \(= \frac{s_{MSW} + s_{SD}}{2} = \textrm{00:59:30}\)

\(\textrm{Min}\) \(= \textrm{Min}_{MSW} - \frac{\overline{X}_{SD}}{2} = \textrm{20:59:00}\)

\(\textrm{Max}\) \(= \textrm{Max}_{MSW} - \frac{\overline{X}_{SD}}{2} = \textrm{01:59:00}\)

Sleep end (\(SE\))

\(\overline{X}\) \(= \overline{X}_{MSW} + \frac{\overline{X}_{SD}}{2} = \textrm{06:51:00}\)

\(s\) \(= \frac{s_{MSW} + s_{SD}}{2} = \textrm{00:59:30}\)

\(\textrm{Min}\) \(= \textrm{Min}_{MSW} + \frac{\overline{X}_{SD}}{2} = \textrm{04:21:00}\)

\(\textrm{Max}\) \(= \textrm{Max}_{MSW} + \frac{\overline{X}_{SD}}{2} = \textrm{09:21:00}\)

Work-free days

Mid-sleep (\(MSW\))

\(\overline{X}\) \(= \textrm{05:02:00}\) (extracted from the base article)

\(s\) \(= \textrm{01:32:00}\) (extracted from the base article)

\(\textrm{Min}\) \(= \overline{X}_{MSW} - (3 \times s_{MSW}) = \textrm{00:26:00}\)

\(\textrm{Max}\) \(= \overline{X}_{MSW} + (3 \times s_{MSW}) = \textrm{09:38:00}\)

Sleep duration (\(SD\))

\(\overline{X}\) \(= \textrm{08:27:00}\) (extracted from the base article)

\(s\) \(= \textrm{01:32:00}\) (extracted from the base article)

\(\textrm{Min}\) \(= \overline{X}_{SD} - (3 \times s_{SD}) = \textrm{03:51:00}\)

\(\textrm{Max}\) \(= \overline{X}_{SD} + (3 \times s_{SD}) = \textrm{13:03:00}\)

Sleep onset (\(SO\))

\(\overline{X}\) \(= \overline{X}_{MSW} - \frac{\overline{X}_{SD}}{2} = \textrm{00:48:30}\)

\(s\) \(= \frac{s_{MSW} + s_{SD}}{2} = \textrm{01:32:00}\)

\(\textrm{Min}\) \(= \textrm{Min}_{MSW} - \frac{\overline{X}_{SD}}{2} = \textrm{20:12:30}\)

\(\textrm{Max}\) \(= \textrm{Max}_{MSW} - \frac{\overline{X}_{SD}}{2} = \textrm{05:24:30}\)

Sleep end (\(SE\))

\(\overline{X}\) \(= \overline{X}_{MSW} + \frac{\overline{X}_{SD}}{2} = \textrm{09:15:30}\)

\(s\) \(= \frac{s_{MSW} + s_{SD}}{2} = \textrm{01:32:00}\)

\(\textrm{Min}\) \(= \textrm{Min}_{MSW} + \frac{\overline{X}_{SD}}{2} = \textrm{04:39:30}\)

\(\textrm{Max}\) \(= \textrm{Max}_{MSW} + \frac{\overline{X}_{SD}}{2} = \textrm{13:51:30}\)

References

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks: daily temporal patterns of human chronotypes. Journal of Biological Rhythms, 18(1), 80-90. doi: 10.1177/0748730402239679 .

Vetter, C., Juda, M., & Roenneberg, T. (2012). The influence of internal time, time awake, and sleep duration on cognitive performance in shiftworkers. Chronobiology international, 29(8), 1127-1138. doi: 10.3109/07420528.2012.707999 .

See also

Examples

if (FALSE) {
random_mctq("standard")
random_mctq("micro")
random_mctq("shift")}