Compute MCTQ overall sleep duration (only for MCTQ Shift)

sd_overall() computes the overall sleep duration for the shift version of the Munich Chronotype Questionnaire (MCTQ).

See sd_week() to compute the average weekly sleep duration for the standard and micro versions of the MCTQ.

sd_overall(n, sd)

Arguments

n
A list object with integerish integer or numeric elements corresponding to the number of days from each shift and number of free days after each shift from a MCTQ shift questionnaire. n elements and values must be paired with sd elements and values.

sd
A list object with Duration elements corresponding to the sleep duration in each shift and sleep duration between two free days after each shift from a MCTQ shift questionnaire (you can use sd() to compute it). sd elements and values must be paired with n elements and values.

Value

A Duration object corresponding to the weighted mean of sd and n (weights).

Details

Standard MCTQ functions were created following the guidelines in Roenneberg, Wirz-Justice, & Merrow (2003), Roenneberg, Allebrandt, Merrow, & Vetter (2012), and The World Wide Experimental Platform (theWeP, n.d.).

\(\mu\)MCTQ functions were created following the guidelines in Ghotbi et.al (2020), in addition to the guidelines used for the standard MCTQ.

MCTQ Shift functions were created following the guidelines in Juda, Vetter, & Roenneberg (2013), in addition to the guidelines used for the standard MCTQ.

See References section to learn more.

Class requirements

The mctq package works with a set of object classes specially created to hold time values. This classes can be found on hms::hms-package and lubridate::lubridate-package. If your data do not conform to the object classes required, you can use convert_to() to convert it (see vignette("converting-data", package = "mctq")).

Rounding and fractional time

Some operations may produce an output with fractional time (e.g. "19538.3828571429s (~5.43 hours)"; 01:15:44.505). If you want, you can round it with round_time().

Our recommendation is to avoid rounding, but, if you do, make sure that you only round your values after all computations are done, that way you can avoid round-off errors.

Operation

The shift version of the MCTQ was developed for shift-workers rotating through morning-, evening-, and night-shifts (transition times at 6:00 a.m., 2:00 p.m., and 10:00 p.m.), but it also allows adaptations to other shift schedules (Juda, Vetter, & Roenneberg, 2013). For that reason, sd_overall() must operate considering any shift combination.

Considering the requirement above, sd_overall() was developed to only accept lists values as arguments. For this approach to work, both n and sd arguments must be lists with paired elements and values between n and sd, i.e. the first element of n (e.g. n_wm) must be paired with the first element of sd (e.g. sd_wm). The function will do the work of combining them and output a weighted mean.

Guidelines

For reference, Juda, Vetter, & Roenneberg (2013) and theWeP (n.d.) guidelines for sd_overall() computation are as follow.

$$((SDwMEN * nwMEN) + (SDfMEN * nfMEN)) / (nwMEN + nfMEN)$$

Where:

\(SDwMEN\) = sleep duration in each shift (MEN*).

\(nwMEN\) = number of days worked in each shift (MEN*) within a shift cycle.

\(SDfMEN\) = sleep duration between two free days after each shift (MEN*).

\(nwMEN\) = number of free days after each shift (MEN*) within a shift cycle.

* \(M\): morning; \(E\): evening; \(N\): night.

Note that the overall sleep duration is the weighted average of the shift-specific sleep durations, i.e. \((SDwMEN * nwMEN)\) and \((SDfMEN * nfMEN)\) must be unfold for each shift and non-shift (e.g. \((SDwM * nw_M) + (SDwE * nwE) ...\)).

References

Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType Questionnaire for shift-workers (MCTQShift). Journal of Biological Rhythms, 28(2), 130-140. doi: 10.1177/0748730412475041.

Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen, D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T. (2020). The \(\mu\)MCTQ: an ultra-short version of the Munich ChronoType Questionnaire. Journal of Biological Rhythms, 35(1), 98-110. doi: 10.1177/0748730419886986.

Roenneberg T., Allebrandt K. V., Merrow M., & Vetter C. (2012). Social jetlag and obesity. Current Biology, 22(10), 939-43. doi: 10.1016/j.cub.2012.03.038.

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks: daily temporal patterns of human chronotypes. Journal of Biological Rhythms, 18(1), 80-90. doi: 10.1177/0748730402239679.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from https://www.thewep.org/documentations/mctq/.

See also

Examples

## __ Scalar example __
n <- list(n_wm = 2, n_we = 4, n_wn = 3, n_fm = 2, n_fw = 4, n_fn = 2)
sd <- list(sd_wm = lubridate::dhours(5), sd_we = lubridate::dhours(6),
           sd_wn = lubridate::dhours(5), sd_fm = lubridate::dhours(10),
           sd_fe = lubridate::dhours(9), sd_fn = lubridate::dhours(8.5))
sd_overall(n, sd)

#> [1] "25835.2941176471s (~7.18 hours)"

#> [1] "25835.2941176471s (~7.18 hours)" # Expected

## __ Vectorized example __
n <- list(n_wm = c(4, 3), n_we = c(2, 3), n_fm = c(2, 2), n_fe = c(4, 4))
sd <- list(sd_wm = c(lubridate::dhours(6), lubridate::dhours(7)),
           sd_we = c(lubridate::dhours(5.5), lubridate::dhours(6)),
           sd_fm = c(lubridate::dhours(9), lubridate::dhours(8.5)),
           sd_fe = c(lubridate::dhours(7), lubridate::dhours(10)))
sd_overall(n, sd)

#> [1] "24300s (~6.75 hours)" "28800s (~8 hours)"   

#> [1] "24300s (~6.75 hours)" "28800s (~8 hours)"  # Expected

## __ Checking the second output from vectorized example __
i <- 2
x <- c(sd[["sd_wm"]][i], sd[["sd_we"]][i], sd[["sd_fm"]][i],
       sd[["sd_fe"]][i])
w <- c(n[["n_wm"]][i], n[["n_we"]][i], n[["n_fm"]][i], n[["n_fe"]][i])
lubridate::as.duration(stats::weighted.mean(x, w))

#> [1] "28800s (~8 hours)"

#> [1] "28800s (~8 hours)" # Expected

## __ Converting the output to hms __
n <- list(n_wm = 3, n_we = 2, n_wn = 2, n_fm = 1, n_fw = 3, n_fn = 5)
sd <- list(sd_wm = lubridate::dhours(4.12), sd_we = lubridate::dhours(6),
           sd_wn = lubridate::dhours(3.43), sd_fm = lubridate::dhours(9.5),
           sd_fe = lubridate::dhours(7.32), sd_fn = lubridate::dhours(10.5))
sd_overall(n, sd)

#> [1] "25915.5s (~7.2 hours)"

#> [1] "25915.5s (~7.2 hours)" # Expected
convert_to(sd_overall(n, sd), "hms")

#> 07:11:55.5

#> 07:11:55.5 # Expected

## __ Rounding the output at the seconds level __
round_time(sd_overall(n, sd))

#> [1] "25916s (~7.2 hours)"

#> [1] "25916s (~7.2 hours)" # Expected
round_time(convert_to(sd_overall(n, sd), "hms"))

#> 07:11:56

#> 07:11:56 # Expected