Compute MCTQ average weekly sleep duration

sd_week() computes the average weekly sleep duration for the standard Munich Chronotype Questionnaire (MCTQ).

sd_week(wd, sd_w, sd_f)

Arguments

wd
A integerish number corresponding to the number of work days per week of a standard MCTQ questionnaire.

sd_w
A Duration vector corresponding to the sleep duration on work days of a standard MCTQ questionnaire (you can use sd() to compute it).

sd_f
A Duration vector corresponding to the sleep duration on work-free days of a standard MCTQ questionnaire (you can use sd() to compute it).

Value

A Duration vector corresponding to the average weekly sleep duration.

Details

This function was created according to Roenneberg, Wirz-Justice & Merrow (2003) and guidelines of The World Wide Experimental Platform (theWeP, n.d.).

Class requirements

The mctq package works with a set of object classes specially created to hold time values. This classes can be found on hms::hms-package and lubridate::lubridate-package. If your data do not conform to the object classes required, you can use convert_to() to convert it.

References

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks: daily temporal patterns of human chronotypes. Journal of Biological Rhythms, 18(1), 80-90. doi: 10.1177/0748730402239679.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from https://www.thewep.org/documentations/mctq/.

See also

Examples

sd_week(5, lubridate::dhours(4), lubridate::dhours(8))

#> [1] "18514.2857142857s (~5.14 hours)"

#> [1] "18514.2857142857s (~5.14 hours)" # Expected
sd_week(4, lubridate::dhours(7), lubridate::dhours(7))

#> [1] "25200s (~7 hours)"

#> [1] "25200s (~7 hours)" # Expected
sd_week(6, lubridate::as.duration(NA), lubridate::dhours(10))

#> [1] NA

#> [1] NA # Expected
x <- c(3, 7)
y <- c(lubridate::dhours(4.5), lubridate::dhours(5.45))
z <- c(lubridate::dhours(8), lubridate::dhours(7.3))
sd_week(x, y, z)

#> [1] "23400s (~6.5 hours)"  "19620s (~5.45 hours)"

#> [1] "23400s (~6.5 hours)"  "19620s (~5.45 hours)" # Expected

## ** converting the output to `hms` **
x <- sd_week(5, lubridate::dhours(5.45), lubridate::dhours(9.5))
convert_to(x, "hms")

#> 06:36:25.714286

#> 06:36:25.714286 # Expected
convert_to(as.integer(x), "hms") # if you want to discard the milliseconds.

#> 06:36:25

#> 06:36:25 # Expected