A fictional MCTQ\(^{Shift}\) dataset

A fictional dataset, for testing and learning purposes, composed by basic/measurable variables of the Munich Chronotype Questionnaire (MCTQ) shift version.

This data was created following the guidelines in Juda, Vetter, & Roenneberg (2013), in addition to the guidelines found in Roenneberg, Wirz-Justice, & Merrow (2003), Roenneberg, Allebrandt, Merrow, & Vetter (2012), and The World Wide Experimental Platform (theWeP, n.d.). See References and Details sections to learn more.

shift_mctq

Format

A tibble with 128 columns and 50 rows:

id

A unique integer value for the purpose of identifying each subject in the dataset.

Type: Control.

R class: integer.

n_w_m

Number of days worked in morning shifts within a shift cycle.

Type: Basic.

R class: integer.

bt_w_m

Local time of going to bed on workdays between two morning shifts.

Statement (EN): "I go to bed at ___ o'clock'".

Type: Basic.

R class: hms.

sprep_w_m

Local time of preparing to sleep on workdays between two morning shifts.

Statement (EN): "I actually get ready to fall asleep at ___ o'clock".

Type: Basic.

R class: hms.

slat_w_m

Sleep latency or time to fall asleep after preparing to sleep on workdays between two morning shifts.

Statement (EN): "I need ___ minutes to fall asleep".

Type: Basic.

R class: Duration.

so_w_m

Local time of sleep onset on workdays between two morning shifts.

Type: Computed.

R class: hms.

se_w_m

Local time of sleep end on workdays between two morning shifts.

Statement (EN): "I wake up at ___ o'clock".

Type: Basic.

R class: hms.

tgu_w_m

Time to get up on workdays between two morning shifts.

Statement (EN): "I get up after ___ minutes".

Type: Basic.

R class: Duration.

gu_w_m

Local time of getting out of bed on workdays between two morning shifts.

Type: Computed.

R class: hms.

alarm_w_m

A logical value indicating if the subject uses an alarm clock to wake up on workdays between two morning shifts.

Statement (EN): "I wake up at ___ o'clock: ( ___ ) with alarm ( ___ ) without alarm".

Type: Basic.

R class: logical.

reasons_w_m

A logical value indicating if the subject have any particular reasons for why he/she cannot freely choose his/her sleep times on workdays between two morning shifts.

Statement (EN): "There are particular reasons why I cannot freely choose my sleep times on morning shifts: Yes ( ___ ) No ( ___ )".

Type: Basic.

R class: logical.

reasons_why_w_m

Particular reasons for why the subject cannot freely choose his/her sleep times on workdays between two morning shifts.

Statement (EN): "If "Yes": Child(ren)/pet(s) ( ___ ) Hobbies ( ___ ) Others, for example: ___".

Type: Basic.

R class: character.

sd_w_m

Sleep duration on workdays between two morning shifts.

Type: Computed.

R class: Duration.

tbt_w_m

Total time in bed on workdays between two morning shifts.

Type: Computed.

R class: Duration.

msw_m

Local time of mid-sleep on workdays between two morning shifts.

Type: Computed.

R class: hms.

nap_w_m

A logical value indicating if the subject usually take a nap on workdays between two morning shifts.

Statement (EN): "I usually take a nap: Yes ( ___ ) No ( ___ )".

Type: Basic.

R class: logical.

napo_w_m

Local time of nap onset on workdays between two morning shifts.

Statement (EN): "If "Yes": I take a nap from ___ o'clock to ___ o'clock".

Type: Basic.

R class: hms.

nape_w_m

Local time of nap end on workdays between two morning shifts.

Statement (EN): "If "Yes": I take a nap from ___ o'clock to ___ o'clock".

Type: Basic.

R class: hms.

napd_w_m

Nap duration on workdays between two morning shifts.

Type: Computed.

R class: Duration.

sd24_w_m

24h sleep duration (sleep duration + nap duration) on workdays between two morning shifts.

Type: Computed.

R class: Duration.

n_f_m

Number of free days after working in morning shifts within a shift cycle.

Type: Basic.

R class: integer.

bt_f_m

Local time of going to bed on work-free days between two free days after morning shifts.

Statement (EN): "I go to bed at ___ o'clock'".

Type: Basic.

R class: hms.

sprep_f_m

Local time of preparing to sleep on work-free days between two free days after morning shifts.

Statement (EN): "I actually get ready to fall asleep at ___ o'clock".

Type: Basic.

R class: hms.

slat_f_m

Sleep latency or time to fall asleep after preparing to sleep on work-free days between two free days after morning shifts.

Statement (EN): "I need ___ minutes to fall asleep".

Type: Basic.

R class: Duration.

so_f_m

Local time of sleep onset on work-free days between two free days after morning shifts.

Type: Computed.

R class: hms.

se_f_m

Local time of sleep end on work-free days between two free days after morning shifts.

Statement (EN): "I wake up at ___ o'clock".

Type: Basic.

R class: hms.

tgu_f_m

Time to get up on work-free days between two free days after morning shifts.

Statement (EN): "I get up after ___ minutes".

Type: Basic.

R class: Duration.

gu_f_m

Local time of getting out of bed on work-free days between two free days after morning shifts.

Type: Computed.

R class: hms.

alarm_f_m

A logical value indicating if the subject uses an alarm clock to wake up on work-free days between two free days after morning shifts.

Statement (EN): "I wake up at ___ o'clock: ( ___ ) with alarm ( ___ ) without alarm".

Type: Basic.

R class: logical.

reasons_f_m

A logical value indicating if the subject have any particular reasons for why he/she cannot freely choose his/her sleep times on work-free days between two free days after morning shifts.

Statement (EN): "There are particular reasons why I cannot freely choose my sleep times on morning shifts: Yes ( ___ ) No ( ___ )".

Type: Basic.

R class: logical.

reasons_why_f_m

Particular reasons for why the subject cannot freely choose his/her sleep times on work-free days between two free days after morning shifts.

Statement (EN): "If "Yes": Child(ren)/pet(s) ( ___ ) Hobbies ( ___ ) Others, for example: ___".

Type: Basic.

R class: character.

sd_f_m

Sleep duration on work-free days between two free days after morning shifts.

Type: Computed.

R class: Duration.

tbt_f_m

Total time in bed on work-free days between two free days after morning shifts.

Type: Computed.

R class: Duration.

msf_m

Local time of mid-sleep on work-free days between two free days after morning shifts.

Type: Computed.

R class: hms.

nap_f_m

A logical value indicating if the subject usually take a nap on work-free days between two free days after morning shifts.

Statement (EN): "I usually take a nap: Yes ( ___ ) No ( ___ )".

Type: Basic.

R class: logical.

napo_f_m

Local time of nap onset on work-free days between two free days after morning shifts.

Statement (EN): "If "Yes": I take a nap from ___ o'clock to ___ o'clock".

Type: Basic.

R class: hms.

nape_f_m

Local time of nap end on work-free days between two free days after morning shifts.

Statement (EN): "If "Yes": I take a nap from ___ o'clock to ___ o'clock".

Type: Basic.

R class: hms.

napd_f_m

Nap duration on work-free days between two free days after morning shifts.

Type: Computed.

R class: Duration.

sd24_f_m

24h sleep duration (sleep duration + nap duration) on work-free days between two free days after morning shifts.

Type: Computed.

R class: Duration.

sd_overall_m

Overall sleep duration considering workdays between two morning shifts and work-free days between two free days after morning shifts.

Type: Computed.

R class: Duration.

msf_sc_m

Corrected local time of mid-sleep on work-free days between two free days after morning shifts.

Type: Computed.

R class: hms.

sjl_rel_m

Relative social jetlag considering workdays between two morning shifts and work-free days between two free days after morning shifts.

Type: Computed.

R class: Duration.

sjl_m

Absolute social jetlag considering workdays between two morning shifts and work-free days between two free days after morning shifts.

Type: Computed.

R class: Duration.

...

For brevity, the subsequent variables, except for sjl_weighted (described below), are not shown here. That's because they have the same configurations of the variables shown above, differing only by shift (evening shift (_e) and night shift (_n)).

sjl_weighted

Absolute social jetlag across all shifts.

Type: Computed.

R class: Duration.

Source

Created by Daniel Vartanian (package's author).

Details

shift_mctq is a tidied, validated, and transformed version of raw_data("shift_mctq.csv").

Guidelines

To learn more about the Munich Chronotype Questionnaire (MCTQ), cf. Roenneberg, Wirz-Justice, & Merrow (2003), Roenneberg, Allebrandt, Merrow, & Vetter (2012), Roenneberg et al. (2015), and Roenneberg, Pilz, Zerbini, & Winnebeck (2019).

To know about different MCTQ versions, cf. Juda, Vetter, & Roenneberg (2013) and Ghotbi et.al (2020).

If you are curious about variable computations and want to have access to the full questionnaire, cf. The Worldwide Experimental Platform (n.d.).

Data building and data wrangling

This dataset was created by randomized sampling (see random_mctq()) and by manual insertions of especial cases. Its purpose is to demonstrate common cases and data issues that researchers may find in their MCTQ data, in addition to be a suggested data structure for MCTQ data.

You can see the shift_mctq build and data wrangling processes at https://github.com/gipsousp/mctq/blob/master/data-raw/shift_mctq.R.

Variable naming

The naming of the variables took into account the naming schemes used in MCTQ articles, in addition to the guidelines of the tidyverse style guide.

Variable classes

The mctq package works with a set of object classes specially created to hold time values. This classes can be found in the hms and lubridate packages.

Duration objects

If you prefer to view Duration objects as hms objects, run pretty_mctq(shift_mctq).

References

Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen, D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T. (2020). The \(\mu\)MCTQ: an ultra-short version of the Munich ChronoType Questionnaire. Journal of Biological Rhythms, 35(1), 98-110. doi: 10.1177/0748730419886986 .

Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType Questionnaire for shift-workers (MCTQ\(^{Shift}\)). Journal of Biological Rhythms, 28(2), 130-140. doi: 10.1177/0748730412475041 .

Roenneberg T., Allebrandt K. V., Merrow M., & Vetter C. (2012). Social jetlag and obesity. Current Biology, 22(10), 939-43. doi: 10.1016/j.cub.2012.03.038 .

Roenneberg, T., Keller, L. K., Fischer, D., Matera, J. L., Vetter, C., & Winnebeck, E. C. (2015). Human activity and rest in situ. In A. Sehgal (Ed.), Methods in enzymology (Vol. 552, pp. 257-283). London, UK: Academic Press. doi: 10.1016/bs.mie.2014.11.028 .

Roenneberg, T., Pilz, L. K., Zerbini, G., & Winnebeck, E. C. (2019). Chronotype and social jetlag: a (self-) critical review. Biology, 8(3), 54. doi: 10.3390/biology8030054 .

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks: daily temporal patterns of human chronotypes. Journal of Biological Rhythms, 18(1), 80-90. doi: 10.1177/0748730402239679 .

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from https://www.thewep.org/documentations/mctq/.

See also