A standard MCTQ dataset

A fictional dataset composed by Munich Chronotype Questionnaire (MCTQ) (standard version) basic/measurable and computed variables for testing and learning purposes.

This data was created according to Roenneberg, Wirz-Justice & Merrow (2003) and guidelines of The World Wide Experimental Platform (theWeP, n.d.). See References and Details sections to learn more.

The naming of the variables took into account the standard names used by the theWeP and guidelines of the tidyverse style guide.

data(std_mctq)

Format

A tibble with 37 columns and 50 rows:

id

A unique numeric value for the purpose of identifying each subject in the dataset.

Type: Control

R class: integer.

work

A logical value indicating if the subject have a regular work schedule.

Statement (EN): "I have a regular work schedule (this includes being, for example, a housewife or househusband)").

Type: Basic

R class: logical.

wd

Number of workdays per week.

Statement (EN): "I have a regular work schedule and work ___ days per week").

Type: Basic.

R class: integer.

fd

Number of work-free days per week.

Type: Computed (7 - wd).

R class: integer.

bt_w

Local time of going to bed on workdays.

Statement (EN): "I go to bed at ___ o'clock'".

Type: Basic.

R class: hms.

sprep_w

Local time of preparing to sleep on workdays.

Statement (EN): "I actually get ready to fall asleep at ___ o'clock".

Type: Basic.

R class: hms.

slat_w

Sleep latency on workdays.

Statement (EN): "I need ___ minutes to fall asleep".

Type: Basic.

R class: Duration.

so_w

Sleep onset on workdays.

Type: Computed (sprep_w + slat_w).

R class: hms.

se_w

Sleep end on workdays.

Statement (EN): "I wake up at ___ o'clock".

Type: Basic.

R class: hms.

si_w

Sleep inertia on workdays.

Statement (EN): "After ___ minutes, I get up".

Type: Basic.

R class: Duration.

gu_w

Local time of getting out of bed on workdays.

Type: Computed (se_w + si_w).

R class: hms.

alarm_w

A logical value indicating if the subject use an alarm clock on workdays.

Statement (EN): "I use an alarm clock on workdays".

Type: Basic.

R class: logical.

wake_before_w

A logical value indicating if the subject regularly wake up BEFORE the alarm rings.

Statement (EN): "I regularly wake up BEFORE the alarm rings".

Type: Basic.

R class: logical.

le_w

Light exposure on workdays.

Statement (EN): "On average, I spend the following amount of time outdoors in daylight (without a roof above my head)".

Type: Basic.

R class: Duration.

sd_w

Sleep duration on workdays.

Type: Computed (se_w - so_w).

R class: Duration.

tbt_w

Total time in bed on workdays.

Type: Computed (gu_w - bt_w).

R class: hms.

msw

Mid-sleep on workdays.

Type: Computed (so_w + (sd_w / 2)).

R class: hms.

bt_f

Local time of going to bed on work-free days.

Statement (EN): "I go to bed at ___ o'clock'".

Type: Basic.

R class: hms.

sprep_f

Local time of preparing to sleep on work-free days

Statement (EN): "I actually get ready to fall asleep at ___ o'clock".

Type: Basic.

R class: hms.

slat_f

Sleep latency on work-free days.

Statement (EN): "I need ___ minutes to fall asleep".

Type: Basic.

R class: Duration.

so_f

Sleep onset on work-free days.

Type: Computed (sprep_f + slat_f).

R class: hms.

se_f

Sleep end on work-free days.

Statement (EN): "I wake up at ___ o'clock".

Type: Basic.

R class: hms.

si_f

Sleep inertia on work-free days.

Statement (EN): "After ___ minutes, I get up".

Type: Basic.

R class: Duration.

gu_f

Local time of getting out of bed on work-free days.

Type: Computed (se_f + si_f).

R class: hms.

alarm_f

A logical value indicating if the subject use an alarm clock on work-free days.

Statement (EN): "My wake-up time is due to the use of an alarm clock".

Type: Basic.

R class: logical.

reasons_f

A logical value indicating if the subject have any particular reasons for why he cannot freely choose his sleep times on work-free days.

Statement (EN): "There are particular reasons why I cannot freely choose my sleep times on free days".

Type: Basic.

R class: logical.

reasons_why_f

Particular reasons why the subject cannot freely choose his sleep times on work-free days.

Statement (EN): "There are particular reasons why I cannot freely choose my sleep times on free days".

Type: Basic.

R class: character.

le_f

Light exposure on work-free days.

Statement (EN): "On average, I spend the following amount of time outdoors in daylight (without a roof above my head)".

Type: Basic.

R class: Duration.

sd_f

Sleep duration on work-free days.

Type: Computed (se_f - so_f).

R class: Duration.

tbt_f

Total time in bed on work-free days.

Type: Computed (gu_f - bt_f).

R class: hms.

msf

Mid-sleep on work-free days.

Type: Computed (so_f + (sd_f / 2)).

R class: hms.

sd_week

Average weekly sleep duration.

Type: Computed.

R class: Duration.

msf_sc

chronotype/corrected midsleep on work-free days.

Type: Computed.

R class: hms.

sloss_week

Weekly sleep loss.

Type: Computed.

R class: Duration.

sjl_rel

Relative social jetlag.

Type: Computed (msf - msw).

R class: Duration.

sjl

Absolute social jetlag.

Type: Computed (abs(msf - msw)).

R class: Duration.

le_week

Average weekly light exposure.

Type: Computed.

R class: Duration.

Source

Prepared by Daniel Vartanian (package's author).

Details

std_mctq is a tidied, validated, and transformed version of mctq:::build_std_mctq(). This dataset was created to demonstrate common cases and data issues that researchers may find in their MCTQ data.

You can learn more about the std_mctq data cleaning process in vignette("data-wrangling", package = "mctq").

To learn more about the Munich Chronotype Questionnaire (MCTQ), cf. Roenneberg, Wirz-Justice & Merrow (2003) and ROENNEBERG, T. et al. (2015).

To know about different MCTQ versions, cf. Juda, Vetter & Roenneberg (2013) and Ghotbi et.al (2020).

If you curious about the computations and want to have access to the full questionnaire, cf. The Worldwide Experimental Platform (n.d.).

References

Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType Questionnaire for shift-workers (MCTQShift). Journal of Biological Rhythms, 28(2), 130-140. doi: 10.1177/0748730412475041.

Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen, D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T. (2020). The \(\mu\)MCTQ: an ultra-short version of the Munich ChronoType Questionnaire. Journal of Biological Rhythms, 35(1), 98-110. doi: 10.1177/0748730419886986.

Roenneberg, T., Keller, L. K., Fischer, D., Matera, J. L., Vetter, C., & Winnebeck, E. C. (2015). Human activity and rest in situ. In A. Sehgal (Ed.), Methods in enzymology (Vol. 552, pp. 257-283). London, UK: Academic Press. doi: 10.1016/bs.mie.2014.11.028.

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks: daily temporal patterns of human chronotypes. Journal of Biological Rhythms, 18(1), 80-90. doi: 10.1177/0748730402239679.

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from https://www.thewep.org/documentations/mctq/.