A fictional standard MCTQ dataset

A fictional dataset composed by standard Munich Chronotype Questionnaire (MCTQ) basic/measurable and computed variables for testing and learning purposes.

This data was created following the guidelines in Roenneberg, Wirz-Justice, & Merrow (2003), Roenneberg, Allebrandt, Merrow, & Vetter (2012), and The World Wide Experimental Platform (theWeP, n.d.). See References and Details sections to learn more.

std_mctq

Format

A tibble with 37 columns and 50 rows:

id

A unique integer value for the purpose of identifying each subject in the dataset.

Type: Control.

R class: integer.

work

A logical value indicating if the subject have a regular work schedule.

Statement (EN): "I have a regular work schedule (this includes being, for example, a housewife or househusband)").

Type: Basic.

R class: logical.

wd

Number of workdays per week.

Statement (EN): "I have a regular work schedule and work ___ days per week".

Type: Basic.

R class: integer.

fd

Number of work-free days per week.

Type: Computed.

R class: integer.

bt_w

Local time of going to bed on workdays.

Statement (EN): "I go to bed at ___ o'clock'".

Type: Basic.

R class: hms.

sprep_w

Local time of preparing to sleep on workdays.

Statement (EN): "I actually get ready to fall asleep at ___ o'clock".

Type: Basic.

R class: hms.

slat_w

Sleep latency on workdays.

Statement (EN): "I need ___ minutes to fall asleep".

Type: Basic.

R class: Duration.

so_w

Sleep onset on workdays.

Type: Computed.

R class: hms.

se_w

Sleep end on workdays.

Statement (EN): "I wake up at ___ o'clock".

Type: Basic.

R class: hms.

si_w

Sleep inertia on workdays.

Statement (EN): "After ___ minutes, I get up".

Type: Basic.

R class: Duration.

gu_w

Local time of getting out of bed on workdays.

Type: Computed.

R class: hms.

alarm_w

A logical value indicating if the subject use an alarm clock on workdays.

Statement (EN): "I use an alarm clock on workdays".

Type: Basic.

R class: logical.

wake_before_w

A logical value indicating if the subject regularly wake up before the alarm rings.

Statement (EN): "I regularly wake up BEFORE the alarm rings".

Type: Basic.

R class: logical.

le_w

Light exposure on workdays.

Statement (EN): "On average, I spend the following amount of time outdoors in daylight (without a roof above my head)".

Type: Extra.

R class: Duration.

sd_w

Sleep duration on workdays.

Type: Computed.

R class: Duration.

tbt_w

Total time in bed on workdays.

Type: Computed.

R class: hms.

msw

Mid-sleep on workdays.

Type: Computed.

R class: hms.

bt_f

Local time of going to bed on work-free days.

Statement (EN): "I go to bed at ___ o'clock'".

Type: Basic.

R class: hms.

sprep_f

Local time of preparing to sleep on work-free days

Statement (EN): "I actually get ready to fall asleep at ___ o'clock".

Type: Basic.

R class: hms.

slat_f

Sleep latency on work-free days.

Statement (EN): "I need ___ minutes to fall asleep".

Type: Basic.

R class: Duration.

so_f

Sleep onset on work-free days.

Type: Computed.

R class: hms.

se_f

Sleep end on work-free days.

Statement (EN): "I wake up at ___ o'clock".

Type: Basic.

R class: hms.

si_f

Sleep inertia on work-free days.

Statement (EN): "After ___ minutes, I get up".

Type: Basic.

R class: Duration.

gu_f

Local time of getting out of bed on work-free days.

Type: Computed.

R class: hms.

alarm_f

A logical value indicating if the subject use an alarm clock on work-free days.

Statement (EN): "My wake-up time is due to the use of an alarm clock".

Type: Basic.

R class: logical.

reasons_f

A logical value indicating if the subject have any particular reasons for why he/she cannot freely choose his/her sleep times on work-free days.

Statement (EN): "There are particular reasons why I cannot freely choose my sleep times on free days".

Type: Basic.

R class: logical.

reasons_why_f

Particular reasons for why the subject cannot freely choose his/her sleep times on work-free days.

Statement (EN): "There are particular reasons why I cannot freely choose my sleep times on free days".

Type: Basic.

R class: character.

le_f

Light exposure on work-free days.

Statement (EN): "On average, I spend the following amount of time outdoors in daylight (without a roof above my head)".

Type: Extra.

R class: Duration.

sd_f

Sleep duration on work-free days.

Type: Computed.

R class: Duration.

tbt_f

Total time in bed on work-free days.

Type: Computed.

R class: hms.

msf

Mid-sleep on work-free days.

Type: Computed.

R class: hms.

sd_week

Average weekly sleep duration.

Type: Computed.

R class: Duration.

msf_sc

Chronotype or corrected mid-sleep on work-free days.

Type: Computed.

R class: hms.

sloss_week

Weekly sleep loss.

Type: Computed.

R class: Duration.

sjl_rel

Relative social jetlag.

Type: Computed.

R class: Duration.

sjl

Absolute social jetlag.

Type: Computed.

R class: Duration.

le_week

Average weekly light exposure.

Type: Computed.

R class: Duration.

Source

Prepared by Daniel Vartanian (package's author).

Details

std_mctq is a tidied, validated, and transformed version of raw_data("std_mctq.csv").

Guidelines

To learn more about the standard Munich Chronotype Questionnaire (MCTQ), cf. Roenneberg, Wirz-Justice, & Merrow (2003), Roenneberg, Allebrandt, Merrow, & Vetter (2012), Roenneberg et al. (2015), and Roenneberg, Pilz, Zerbini, & Winnebeck (2019).

To know about different MCTQ versions, cf. Juda, Vetter, & Roenneberg (2013) and Ghotbi et.al (2020).

If you curious about the variable computations and want to have access to the full questionnaire, cf. The Worldwide Experimental Platform (n.d.).

Data building and data wrangling

This dataset was created by randomized sampling processes and by manual insertions of especial cases. Its purpose is to demonstrate common cases and data issues that researchers may find in their MCTQ data, in addition to be a suggested data structure for MCTQ data.

You can see the std_mctq build and data wrangling processes at https://bit.ly/3cvXxKL.

Variable naming

The naming of the variables took into account the standard names/abbreviations used by Roenneberg, Allebrandt, Merrow, & Vetter (2012) and theWeP (n.d.), in addition to the guidelines of the tidyverse style guide.

Variable classes

The mctq package works with a set of object classes specially created to hold time values. This classes can be found on the hms::hms-package and lubridate::lubridate-package.

References

Ghotbi, N., Pilz, L. K., Winnebeck, E. C., Vetter, C., Zerbini, G., Lenssen, D., Frighetto, G., Salamanca, M., Costa, R., Montagnese, S., & Roenneberg, T. (2020). The \(\mu\)MCTQ: an ultra-short version of the Munich ChronoType Questionnaire. Journal of Biological Rhythms, 35(1), 98-110. doi: 10.1177/0748730419886986 .

Juda, M., Vetter, C., & Roenneberg, T. (2013). The Munich ChronoType Questionnaire for shift-workers (MCTQ\(^{Shift}\)). Journal of Biological Rhythms, 28(2), 130-140. doi: 10.1177/0748730412475041 .

Roenneberg T., Allebrandt K. V., Merrow M., & Vetter C. (2012). Social jetlag and obesity. Current Biology, 22(10), 939-43. doi: 10.1016/j.cub.2012.03.038 .

Roenneberg, T., Keller, L. K., Fischer, D., Matera, J. L., Vetter, C., & Winnebeck, E. C. (2015). Human activity and rest in situ. In A. Sehgal (Ed.), Methods in enzymology (Vol. 552, pp. 257-283). London, UK: Academic Press. doi: 10.1016/bs.mie.2014.11.028 .

Roenneberg, T., Pilz, L. K., Zerbini, G., & Winnebeck, E. C. (2019). Chronotype and social jetlag: a (self-) critical review. Biology, 8(3), 54. doi: 10.3390/biology8030054 .

Roenneberg, T., Wirz-Justice, A., & Merrow, M. (2003). Life between clocks: daily temporal patterns of human chronotypes. Journal of Biological Rhythms, 18(1), 80-90. doi: 10.1177/0748730402239679 .

The Worldwide Experimental Platform (n.d.). MCTQ. Retrieved from https://www.thewep.org/documentations/mctq/.